refactor(account-overview): split derived-advisories engine into per-detector classes
Break the 2288-line derivedAdvisories.ts into a modular advisories/ package:
one class + file per detector (18), all extending a shared Detector base class,
a registry-driven engine with declarative per-scope suppression, and separated
metrics/logs fetchers under fetchers/. Pure code move - no logic, threshold,
or string changes; behavior verified by the unchanged unit + parity tests
(301/301 passing). Also breaks the derivedAdvisories<->logs circular import by
extracting shared types into core/types.
