--- src/panels/accountOverview/services/derivedAdvisories.ts
@@ -1,2288 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { type CosmosDBManagementClient } from '@azure/arm-cosmosdb';
-import { type MonitorClient } from '@azure/arm-monitor';
-import * as l10n from '@vscode/l10n';
-import {
-    fetchCrossPartitionShapes,
-    fetchSharedThroughputTraffic,
-    fetchUncontrolledIngestion,
-    logWindow,
-    probeCdbLogs,
-    type LogsQueryExecutor,
-} from './derivedAdvisoryLogs';
-import {
-    getAccountConsumedRuShape,
-    getAutoscaleUtilization,
-    getContainerIdlePeaks,
-    getContainerPartitionCounts,
-} from './derivedAdvisoryMetrics';
-import { classifyEnvironment } from './environmentClassifier';
-import { getSqlInventory, type ThroughputMode } from './inventory';
-import { getInventoryMetrics, type HealthThresholds } from './inventoryMetrics';
-import {
-    getPartitionHealth,
-    getPartitionStorageSeries,
-    type PartitionStorageResult,
-    type PartitionStorageSample,
-    type PartitionStorageSeries,
-    type PartitionThresholds,
-} from './partitionHealth';
-import { getRuTrends, getThrottleRate } from './ruTrends';
-import {
-    containerKey,
-    DEFAULT_PARTITION_HEADROOM_PCT,
-    DEFAULT_PARTITION_SATURATION_PCT,
-    DEFAULT_THROTTLE_RATE_PCT,
-    percentile,
-    type UnavailableReason,
-} from './shared';
-
-// ─── Client-side derived advisories ──────────────────────────────────────────────
-//
-// Unlike Azure Alerts + Advisor recommendations, these advisories are derived on
-// the extension host from telemetry the dashboard has already fetched (RU trends,
-// inventory metrics, partition health). No new ARM endpoints. The portal does not
-// derive these on its Overview blade, so thresholds are advisory and user-tunable
-// via `cosmosDB.accountOverview.advisories.*`. Every rule function is pure so it
-// can be unit-tested against synthetic inputs; `collectDerivedAdvisories` does the
-// fetching and calls `computeDerivedAdvisories`.
-
-/** The derived-advisory rules this engine implements (P0 + best-effort). */
-export type DerivedAdvisoryRule =
-    | 'HotPartitionRisk'
-    | 'SustainedThrottlingInRegion'
-    | 'OverProvisioning'
-    | 'AutoscaleCandidate'
-    | 'StorageGrowthRisk'
-    | 'StorageSkewRisk'
-    | 'IndexingCostRisk'
-    | 'ExpensiveConsistency'
-    | 'MultiRegionWriteAntipattern'
-    | 'IdleContainer'
-    | 'PartitionMergeCandidate'
-    | 'AutoscaleMaxOverProvisioned'
-    | 'AutoscaleToManualCandidate'
-    | 'ServerlessCandidate'
-    | 'CrossPartitionQuery'
-    | 'ShardKeyMisalignment'
-    | 'UncontrolledIngestion'
-    | 'SharedThroughputStarvation';
-
-export type DerivedAdvisorySeverity = 'High' | 'Medium' | 'Low';
-
-export interface DerivedAdvisory {
-    /** Stable id (rule + optional scope) so the webview can dismiss it for the session. */
-    id: string;
-    rule: DerivedAdvisoryRule;
-    severity: DerivedAdvisorySeverity;
-    title: string;
-    /** Why this fired; ≤ 500 chars. */
-    rationale: string;
-    suggestedAction: string;
-    /** Human-readable reference to the configured threshold that fired the rule. */
-    thresholdReference: string;
-    /** `databaseId/containerId` this advisory concerns, when container-scoped. */
-    scope?: string;
-}
-
-/**
- * Per-tier availability of the Log Analytics ("Tier-2") data path, so the card can render the Tier-1
- * advisories that *did* run while flagging that the log-based analyzers were skipped. This is the
- * partial-coverage contract: Tier-2 degrading must never blank the whole card (which always has
- * Tier-1 metrics + config advisories to show).
- */
-export interface LogsSourceStatus {
-    /** True when the `CDB*` tables were queryable (diagnostic settings on + Reader RBAC), even if empty. */
-    available: boolean;
-    /** When `available` is false, why the log-based (Tier-2) analyzers were skipped. */
-    reason?: UnavailableReason;
-}
-
-export interface DerivedAdvisoriesResult {
-    /** False when none of the underlying telemetry was available to evaluate. */
-    available: boolean;
-    /** When `available` is false, why: `noData` | `unsupported` | `rbac` | `logAnalyticsDisabled`. */
-    reason?: UnavailableReason;
-    advisories: DerivedAdvisory[];
-    /**
-     * Availability of the Log Analytics ("Tier-2") analyzers, present when the section itself is available.
-     * Absent (or `available: true`) means the log-based checks ran; `available: false` drives the card's
-     * partial-coverage notice while the Tier-1 advisories above still render.
-     */
-    logSource?: LogsSourceStatus;
-    /** Epoch milliseconds when the server produced this snapshot. */
-    generatedAt: number;
-}
-
-/** Thresholds for the derived-advisory rules, sourced from `cosmosDB.accountOverview.*`. */
-export interface DerivedAdvisoryThresholds {
-    /**
-     * RU: a physical partition's p99 (% of provisioned) at or above which it counts as saturated. Drives both the
-     * HotPartitionRisk (DX-006) and SustainedThrottlingInRegion (DX-005) rules — a container is only flagged when a
-     * partition is actually at capacity.
-     */
-    partitionSaturationPercent: number;
-    /**
-     * RU: a physical partition's p99 below which it still has headroom. A saturated busiest partition *with* a
-     * cooler sibling below this is a hot partition (DX-006, re-key); every partition at/above this is uniform
-     * under-provisioning (DX-005, add RU/s).
-     */
-    partitionHeadroomPercent: number;
-    /** Container 429 rate (`sum(429)/sum(total)`, %) at or above which throttling counts as active. */
-    throttleRatePercent: number;
-    /** 7-day p99 RU (%) below which OverProvisioning fires (CODA DX-001 moderate band). */
-    overProvisioningBandPercent: number;
-    /** Peak RU (%) a workload must reach for AutoscaleCandidate — below this you should right-size, not autoscale. */
-    autoscaleMaxPercent: number;
-    /** Average RU (%) a workload must stay at or below for AutoscaleCandidate (mostly idle between bursts). */
-    autoscaleAvgPercent: number;
-    /** Peak-to-average RU ratio at or above which AutoscaleCandidate fires (a genuine burst, not steady load). */
-    autoscalePeakToAvgRatio: number;
-    /**
-     * Projected days-to-limit horizon at or below which StorageGrowthRisk fires. A physical partition whose
-     * least-squares storage trajectory reaches the 50 GiB split ceiling within this many days is flagged; severity
-     * grades by how soon (≤ 30 d High, ≤ 90 d Medium, otherwise Low).
-     */
-    storageGrowthHorizonDays: number;
-    /**
-     * Balance ratio (coolest physical partition size ÷ busiest) below which, when the busiest partition is also
-     * material, StorageSkewRisk fires. 1.0 is perfectly balanced; a low value means one partition holds far more
-     * than its siblings and will hit the split ceiling long before they do.
-     */
-    storageSkewBalanceRatio: number;
-    /** Index/data storage ratio above which IndexingCostRisk fires. */
-    indexingUsageRatio: number;
-    /**
-     * RU: peak RU consumed in any single bucket (max of `TotalRequestUnits`, Total) at or below which a container
-     * counts as idle (DX-004). A container that never spends more than this near-zero floor in any bucket over the
-     * 30-day window did no real work — its whole reservation (manual) or idle floor (autoscale) is recoverable.
-     */
-    idlePeakRuPerBucket: number;
-    /**
-     * Autoscale utilisation (average % of the configured max) at or above which AutoscaleToManualCandidate (DX-013)
-     * fires — a workload sustained near its max barely uses autoscale's elasticity and is cheaper on manual. Grounded
-     * in Azure's autoscale break-even: autoscale only saves money if `Tmax` is used 66% or fewer hours per month.
-     * https://learn.microsoft.com/azure/cosmos-db/provision-throughput-autoscale#benefits-of-autoscale
-     */
-    autoscaleToManualAvgPercent: number;
-    /**
-     * Peak-to-average ratio of autoscale utilisation at or below which the load counts as steady (DX-013). Combined
-     * with the average floor, a flat high series (little spikiness) is the autoscale→manual signal.
-     */
-    autoscaleToManualPeakToAvgRatio: number;
-    /**
-     * Account-total average-to-peak RU/s ratio below which ServerlessCandidate (DX-014) fires — a low, sporadic
-     * workload with long idle stretches whose shape suits serverless pay-per-RU billing.
-     */
-    serverlessSporadicRatio: number;
-    /**
-     * RU/s: account-total peak above which ServerlessCandidate (DX-014) considers the workload real enough to
-     * evaluate (below this the account is idle — that is DX-004's decommission case, not serverless).
-     */
-    serverlessPeakFloorRuPerSec: number;
-    /**
-     * RU/s: account-total peak at or below which ServerlessCandidate (DX-014) still fits serverless. Above this
-     * single-partition ceiling the workload is steered away — a serverless container caps at 5,000 RU/s:
-     * https://learn.microsoft.com/azure/cosmos-db/serverless-performance#request-unit-changes
-     */
-    serverlessPeakCeilingRuPerSec: number;
-    // ─── Tier-2 (Log Analytics) thresholds ─────────────────────────────────────
-    /** DX-002: minimum query executions in the window below which there is too little traffic to judge fan-out. */
-    crossPartitionMinQueries: number;
-    /** DX-002: a query shape whose average fan-out (distinct partitions per logical query) is at or above this counts as cross-partition. */
-    crossPartitionFanoutThreshold: number;
-    /** DX-002: share (%) of executions fanning out at or above which CrossPartitionQuery is High (else Medium down to the med floor). */
-    crossPartitionHighPct: number;
-    /** DX-002: share (%) of executions fanning out below which CrossPartitionQuery does not fire. */
-    crossPartitionMedPct: number;
-    /** DX-007: share (%) of executions fanning out at or above which the partition *key* is structurally misaligned (re-key), superseding DX-002. */
-    shardKeyStructuralPct: number;
-    /** DX-007: cross-partition share (%) at or above which ShardKeyMisalignment is High (else Medium). */
-    shardKeyHighPct: number;
-    /** DX-010: write-op RU share (%) at or above which the workload is write-dominant. */
-    ingestionWriteRuPctFloor: number;
-    /** DX-010: 429 rate (%) at or above which the container is throttling (paired with write-dominance to fire). */
-    ingestionThrottleRatePct: number;
-    /** DX-010: 429 rate (%) at or above which UncontrolledIngestion is High (else Medium). */
-    ingestionHighPct: number;
-    /** DX-010: minimum data-plane requests in the window below which there is too little traffic to judge. */
-    ingestionMinRequests: number;
-    /** DX-003: a collection with fewer requests than this is inactive for the window and ignored. */
-    sharedThroughputMinRequests: number;
-    /** DX-003: database-wide 429 rate (%) at or above which the shared pool is under pressure. */
-    sharedThroughputPoolThrottlePct: number;
-    /** DX-003: RU share (%) of the pool one collection must consume to count as the monopolizer. */
-    sharedThroughputDominancePct: number;
-    /** DX-003: 429 rate (%) at or above which a low-consumption sibling counts as a starved victim. */
-    sharedThroughputVictimThrottlePct: number;
-    /** DX-003: RU share (%) at or below which a throttled sibling counts as a starved victim (consuming little). */
-    sharedThroughputVictimSharePct: number;
-}
-
-export const DEFAULT_ADVISORY_THRESHOLDS: DerivedAdvisoryThresholds = {
-    partitionSaturationPercent: DEFAULT_PARTITION_SATURATION_PCT,
-    partitionHeadroomPercent: DEFAULT_PARTITION_HEADROOM_PCT,
-    throttleRatePercent: DEFAULT_THROTTLE_RATE_PCT,
-    overProvisioningBandPercent: 30,
-    autoscaleMaxPercent: 40,
-    autoscaleAvgPercent: 30,
-    autoscalePeakToAvgRatio: 5,
-    storageGrowthHorizonDays: 180,
-    storageSkewBalanceRatio: 0.7,
-    indexingUsageRatio: 0.3,
-    idlePeakRuPerBucket: 50,
-    autoscaleToManualAvgPercent: 66,
-    autoscaleToManualPeakToAvgRatio: 1.3,
-    serverlessSporadicRatio: 0.1,
-    serverlessPeakFloorRuPerSec: 10,
-    serverlessPeakCeilingRuPerSec: 5000,
-    // Tier-2 (Log Analytics) — CODA-calibrated (cross_partition_query.py, shard_key_misalignment.py,
-    // uncontrolled_ingestion.py, shared_throughput.py). Internal-guidance defaults, tunable via advisories.*.
-    crossPartitionMinQueries: 50,
-    crossPartitionFanoutThreshold: 1.5,
-    crossPartitionHighPct: 30,
-    crossPartitionMedPct: 10,
-    shardKeyStructuralPct: 60,
-    shardKeyHighPct: 80,
-    ingestionWriteRuPctFloor: 80,
-    ingestionThrottleRatePct: 10,
-    ingestionHighPct: 20,
-    ingestionMinRequests: 1000,
-    sharedThroughputMinRequests: 100,
-    sharedThroughputPoolThrottlePct: 5,
-    sharedThroughputDominancePct: 60,
-    sharedThroughputVictimThrottlePct: 5,
-    sharedThroughputVictimSharePct: 20,
-};
-
-const ADVISORY_SEVERITY_ORDER: Record<DerivedAdvisorySeverity, number> = { High: 0, Medium: 1, Low: 2 };
-
-/** Sorts advisories by severity (High first), then rule, then scope. Pure. */
-export function compareAdvisories(a: DerivedAdvisory, b: DerivedAdvisory): number {
-    return (
-        ADVISORY_SEVERITY_ORDER[a.severity] - ADVISORY_SEVERITY_ORDER[b.severity] ||
-        a.rule.localeCompare(b.rule) ||
-        (a.scope ?? '').localeCompare(b.scope ?? '')
-    );
-}
-
-/** Defensive clamp so a rationale never exceeds the 500-char budget. */
-function clampRationale(text: string): string {
-    return text.length <= 500 ? text : text.slice(0, 499) + '…';
-}
-
-/** Arithmetic mean of the finite values in a series; 0 for an empty series. Pure. */
-export function mean(values: readonly number[]): number {
-    const clean = values.filter((v) => Number.isFinite(v));
-    if (clean.length === 0) {
-        return 0;
-    }
-    return clean.reduce((sum, v) => sum + v, 0) / clean.length;
-}
-
-// ─── Shared capacity-materiality severity (CODA framework/materiality) ─────────────
-//
-// The cost-oriented rules (DX-001 over-provisioning, DX-004 idle, DX-011 autoscale-max) grade severity by
-// **relative materiality in capacity terms**: a finding's recoverable/wasted RU/s as a percentage of the scope's
-// total provisioned RU/s (the same waste is material against a small account, trivial against a fleet). When the
-// scope total is unknown the model falls back to absolute-RU bands. CODA reports capacity, not cost (ADR-0012).
-
-/** Recoverable RU/s as % of scope provisioned RU/s at or above which severity is High. */
-const MATERIALITY_HIGH_PCT = 5;
-/** Recoverable RU/s as % of scope provisioned RU/s at or above which severity is Medium (below this, Low). */
-const MATERIALITY_MEDIUM_PCT = 1;
-/** Absolute recoverable RU/s at or above which severity is High, used only when the scope total is unknown. */
-const MATERIALITY_ABS_HIGH_RU = 5000;
-/** Absolute recoverable RU/s at or above which severity is Medium, used only when the scope total is unknown. */
-const MATERIALITY_ABS_MEDIUM_RU = 1000;
-
-/**
- * Maps a finding's RU/s (recoverable for idle, wasted for over-provisioning) to a severity — relative to the
- * scope's provisioned RU/s when known, else an absolute-RU fallback. Shared so DX-001/DX-004/DX-011 stay
- * consistent (CODA framework/materiality, ADR-0012). Pure.
- */
-function capacityMaterialitySeverity(
-    findingRu: number,
-    scopeProvisionedRu: number | undefined,
-): DerivedAdvisorySeverity {
-    if (scopeProvisionedRu !== undefined && scopeProvisionedRu > 0) {
-        const pct = (findingRu / scopeProvisionedRu) * 100;
-        if (pct >= MATERIALITY_HIGH_PCT) {
-            return 'High';
-        }
-        return pct >= MATERIALITY_MEDIUM_PCT ? 'Medium' : 'Low';
-    }
-    if (findingRu >= MATERIALITY_ABS_HIGH_RU) {
-        return 'High';
-    }
-    return findingRu >= MATERIALITY_ABS_MEDIUM_RU ? 'Medium' : 'Low';
-}
-
-/**
- * Already-fetched per-container saturation picture the hot-partition (DX-006) and under-provisioning (DX-005)
- * rules consume. `maxP99`/`minP99` are the busiest/coolest physical partition p99 of `NormalizedRUConsumption`
- * over the window; `throttleRatePercent` is the container's aggregate 429 rate.
- */
-export interface PartitionSaturationInput {
-    databaseId: string;
-    containerId: string;
-    /** Busiest physical partition's p99 utilization (% of provisioned). */
-    maxP99: number;
-    /** Coolest physical partition's p99 utilization (% of provisioned). */
-    minP99: number;
-    partitionCount: number;
-    /** Container 429 rate (`sum(429)/sum(total)`, %). */
-    throttleRatePercent: number;
-    /** Total requests over the window; below the abstain gate the rate is too noisy to judge. */
-    totalRequests: number;
-    /** Throughput mode, so the under-provisioning remediation can say "raise the max" vs "raise RU/s". */
-    throughputMode: ThroughputMode;
-    /** Provisioned RU/s, quoted in the under-provisioning remediation when known. */
-    provisionedRu?: number;
-}
-
-/**
- * HotPartitionRisk (DX-006): the busiest physical partition is **saturated** (p99 at or above the saturation
- * threshold) while at least one other partition still has **headroom** (p99 below the headroom threshold). Unlike a
- * raw share or fair-share cutoff, this keys off "is some partition pinned at capacity while another is cool" — the
- * signal that the partition key, not more RU/s, is the fix. A single partition, or uniform saturation (every
- * partition busy → global under-provisioning, DX-005), never fires here. Severity scales with actual impact: active
- * throttling (429 rate at or above the threshold) → High, saturated-but-not-throttling → Medium. Pure.
- */
-export function evaluateHotPartitionRisk(
-    input: PartitionSaturationInput,
-    saturationPercent: number,
-    headroomPercent: number,
-    throttleThresholdPercent: number,
-): DerivedAdvisory | undefined {
-    if (input.partitionCount < 2 || input.maxP99 < saturationPercent || input.minP99 >= headroomPercent) {
-        return undefined;
-    }
-    const scope = containerKey(input.databaseId, input.containerId);
-    const busiest = Math.round(input.maxP99);
-    const coolest = Math.round(input.minP99);
-    const saturation = Math.round(saturationPercent);
-    const headroom = Math.round(headroomPercent);
-    const throttling = input.throttleRatePercent >= throttleThresholdPercent;
-    const severity: DerivedAdvisorySeverity = throttling ? 'High' : 'Medium';
-
-    const impact = throttling
-        ? l10n.t('The container is already throttling (HTTP 429), so the hot partition is capping real traffic.')
-        : l10n.t(
-              'It is not throttling yet, but the pinned partition cannot borrow the cooler partitions’ capacity as load grows.',
-          );
-
-    return {
-        id: `HotPartitionRisk:${scope}`,
-        rule: 'HotPartitionRisk',
-        severity,
-        title: l10n.t('Hot physical partition in {container}', { container: input.containerId }),
-        rationale: clampRationale(
-            l10n.t(
-                'In "{container}", the busiest of {count} physical partitions ran at {busiest}% p99 utilization (at or above the {saturation}% saturation mark) while the coolest sat at {coolest}% (below the {headroom}% headroom mark). Load is concentrated by the partition key, not a global shortfall. ',
-                { container: input.containerId, count: input.partitionCount, busiest, saturation, coolest, headroom },
-            ) + impact,
-        ),
-        suggestedAction: l10n.t(
-            'Redesign the partition key for higher cardinality (avoid hotspots; consider synthetic or hierarchical keys) so traffic spreads across partitions — adding RU/s will not fix a single hot partition.',
-        ),
-        thresholdReference: l10n.t('Threshold: busiest partition p99 ≥ {saturation}% while another < {headroom}%', {
-            saturation,
-            headroom,
-        }),
-        scope,
-    };
-}
-
-/** Container 429 rate (%) at or above which under-provisioning severity is High. */
-const UNDER_PROVISIONING_HIGH_RATE_PCT = 20;
-/** Container 429 rate (%) at or above which under-provisioning severity is Medium (below this, Low). */
-const UNDER_PROVISIONING_MEDIUM_RATE_PCT = 5;
-/** Below this request count the 429 rate is too noisy to judge, so the rule abstains. */
-const UNDER_PROVISIONING_MIN_REQUESTS = 1_000;
-
-/**
- * SustainedThrottlingInRegion / genuine under-provisioning (DX-005): a container is sustaining 429 throttling
- * (rate at or above the threshold) **and** at least one partition is saturated, **and** the saturation is
- * *uniform* — every partition is busy (coolest p99 at or above the headroom mark). Uniform saturation means the
- * workload simply needs more capacity; the non-uniform (skew) case is left to HotPartitionRisk (DX-006) so the two
- * never double-report the same throttling. A `min-requests` gate abstains on traffic too low to judge. Severity
- * grades by the 429 rate. Pure.
- */
-export function evaluateUnderProvisioning(
-    input: PartitionSaturationInput,
-    saturationPercent: number,
-    headroomPercent: number,
-    throttleThresholdPercent: number,
-): DerivedAdvisory | undefined {
-    if (input.totalRequests < UNDER_PROVISIONING_MIN_REQUESTS) {
-        return undefined;
-    }
-    const throttling = input.throttleRatePercent >= throttleThresholdPercent;
-    const atCapacity = input.maxP99 >= saturationPercent;
-    const uniform = input.minP99 >= headroomPercent;
-    if (!throttling || !atCapacity || !uniform) {
-        return undefined;
-    }
-    const scope = containerKey(input.databaseId, input.containerId);
-    const rate = Math.round(input.throttleRatePercent * 10) / 10;
-    const busiest = Math.round(input.maxP99);
-    const coolest = Math.round(input.minP99);
-    const severity: DerivedAdvisorySeverity =
-        input.throttleRatePercent >= UNDER_PROVISIONING_HIGH_RATE_PCT
-            ? 'High'
-            : input.throttleRatePercent >= UNDER_PROVISIONING_MEDIUM_RATE_PCT
-              ? 'Medium'
-              : 'Low';
-
-    const action =
-        input.throughputMode === 'autoscale'
-            ? input.provisionedRu !== undefined
-                ? l10n.t('Raise the autoscale maximum above {ru} RU/s — it is pinned at the ceiling.', {
-                      ru: Math.round(input.provisionedRu).toLocaleString(),
-                  })
-                : l10n.t('Raise the autoscale maximum — it is pinned at the ceiling.')
-            : input.provisionedRu !== undefined
-              ? l10n.t('Raise provisioned throughput above {ru} RU/s, or convert to autoscale.', {
-                    ru: Math.round(input.provisionedRu).toLocaleString(),
-                })
-              : l10n.t('Raise provisioned throughput, or convert to autoscale.');
-
-    return {
-        id: `SustainedThrottlingInRegion:${scope}`,
-        rule: 'SustainedThrottlingInRegion',
-        severity,
-        title: l10n.t('Sustained throttling in {container}', { container: input.containerId }),
-        rationale: clampRationale(
-            l10n.t(
-                'In "{container}", {rate}% of requests were throttled (HTTP 429) while every physical partition was saturated (busiest p99 {busiest}%, coolest {coolest}%). Because all partitions are at capacity — not just one — this is a genuine capacity shortfall rather than a hot partition.',
-                { container: input.containerId, rate, busiest, coolest },
-            ),
-        ),
-        suggestedAction: action,
-        thresholdReference: l10n.t('Threshold: 429 rate ≥ {rate}% with every partition p99 ≥ {saturation}%', {
-            rate: Math.round(throttleThresholdPercent),
-            saturation: Math.round(saturationPercent),
-        }),
-        scope,
-    };
-}
-
-/** Do not flag over-provisioning when the busiest sample reached this % — a recurring batch needs peak capacity. */
-const OVERPROVISIONING_PEAK_GUARD_PCT = 90;
-/** Headroom multiplier applied to observed p99 demand when sizing the right-size target (CODA DX-001). */
-const OVERPROVISIONING_HEADROOM = 1.3;
-// Platform minimum RU/s floor a right-size recommendation never drops below — a container's minimum manual
-// throughput is 400 RU/s:
-// https://learn.microsoft.com/azure/cosmos-db/set-throughput
-const OVERPROVISIONING_MIN_RU = 400;
-/** Wasted RU/s as % of scope provisioned RU/s at or above which severity is High. */
-const OVERPROVISIONING_MATERIAL_HIGH_PCT = 5;
-/** Wasted RU/s as % of scope provisioned RU/s at or above which severity is Medium (below this, Low). */
-const OVERPROVISIONING_MATERIAL_MEDIUM_PCT = 1;
-
-/**
- * OverProvisioning (CODA DX-001): sustained low **p99** RU on manual throughput over the trailing 7 days, banded
- * below the moderate threshold. Two guards keep it from firing on workloads that genuinely need their capacity: the
- * p99 band ignores brief spikes, and a **peak-saturation guard** suppresses the finding when the busiest sample
- * saturated capacity (a recurring batch needs that peak). Severity is **relative materiality** — wasted RU/s as a
- * share of the scope's provisioned RU/s (the same waste is material against a small account but trivial against a
- * fleet) — falling back to absolute wasted RU/s when the scope total is unknown. The right-size target covers p99
- * demand plus headroom and never drops below the observed peak, so it can never throttle a real spike. Pure.
- */
-export function evaluateOverProvisioning(
-    p99Percent: number | undefined,
-    peakPercent: number | undefined,
-    thresholdPercent: number,
-    hasManualThroughput: boolean,
-    provisionedRuTotal?: number,
-): DerivedAdvisory | undefined {
-    if (!hasManualThroughput || p99Percent === undefined || p99Percent >= thresholdPercent) {
-        return undefined;
-    }
-    // Peak-saturation guard: a workload whose busiest sample saturates capacity genuinely needs it at peak.
-    if (peakPercent !== undefined && peakPercent >= OVERPROVISIONING_PEAK_GUARD_PCT) {
-        return undefined;
-    }
-    const p99 = Math.round(p99Percent);
-    const threshold = Math.round(thresholdPercent);
-
-    let wastedRu: number | undefined;
-    let severity: DerivedAdvisorySeverity = 'Medium';
-    if (provisionedRuTotal !== undefined && provisionedRuTotal > 0) {
-        const consumedAtP99 = (p99Percent / 100) * provisionedRuTotal;
-        const peakConsumed = ((peakPercent ?? p99Percent) / 100) * provisionedRuTotal;
-        const recommended = Math.min(
-            provisionedRuTotal,
-            Math.max(
-                OVERPROVISIONING_MIN_RU,
-                Math.ceil(consumedAtP99 * OVERPROVISIONING_HEADROOM),
-                Math.ceil(peakConsumed),
-            ),
-        );
-        wastedRu = Math.max(0, Math.round(provisionedRuTotal - recommended));
-        const materialityPct = (wastedRu / provisionedRuTotal) * 100;
-        severity =
-            materialityPct >= OVERPROVISIONING_MATERIAL_HIGH_PCT
-                ? 'High'
-                : materialityPct >= OVERPROVISIONING_MATERIAL_MEDIUM_PCT
-                  ? 'Medium'
-                  : 'Low';
-    }
-
-    const materiality =
-        wastedRu !== undefined
-            ? l10n.t(' That is roughly {wasted} RU/s of provisioned capacity the workload never uses.', {
-                  wasted: wastedRu.toLocaleString(),
-              })
-            : '';
-
-    return {
-        id: 'OverProvisioning',
-        rule: 'OverProvisioning',
-        severity,
-        title: l10n.t('Throughput may be over-provisioned'),
-        rationale: clampRationale(
-            l10n.t(
-                '99th-percentile normalized RU consumption stayed at {p99}% over the last 7 days, below the {threshold}% over-provisioning threshold, while at least one container uses manual throughput. Banding on p99 rather than the peak ignores brief spikes that do not reflect steady demand.',
-                { p99, threshold },
-            ) + materiality,
-        ),
-        suggestedAction: l10n.t(
-            'Lower provisioned RU/s to match observed demand, or switch to autoscale so capacity tracks usage automatically.',
-        ),
-        thresholdReference: l10n.t('Threshold: 7-day p99 < {threshold}% RU', { threshold }),
-    };
-}
-
-/** Duty-cycle thresholds for {@link evaluateAutoscaleCandidate}. */
-export interface AutoscaleThresholds {
-    /** Peak RU (%) the workload must reach — below this, right-size instead of autoscaling. */
-    maxPercent: number;
-    /** Average RU (%) the workload must stay at or below (mostly idle between bursts). */
-    avgPercent: number;
-    /** Peak-to-average ratio at or above which the burst is genuine. */
-    peakToAvgRatio: number;
-}
-
-/**
- * AutoscaleCandidate (best-effort): a manual container whose 7-day RU profile is
- * a genuine burst — a real peak (`max ≥ maxPercent`) on a mostly-idle baseline
- * (`avg ≤ avgPercent`) with a large peak-to-average ratio (`≥ peakToAvgRatio`).
- * This is the duty cycle (average as a fraction of peak) that autoscale
- * economics turn on, not dispersion around the mean — a workload oscillating
- * 40↔60% has high variance yet a high duty cycle, so autoscale would cost more.
- * Pure.
- */
-export function evaluateAutoscaleCandidate(
-    maxPercent: number | undefined,
-    avgPercent: number | undefined,
-    thresholds: AutoscaleThresholds,
-    hasManualThroughput: boolean,
-): DerivedAdvisory | undefined {
-    if (!hasManualThroughput || maxPercent === undefined || avgPercent === undefined || avgPercent <= 0) {
-        return undefined;
-    }
-    const peakToAvg = maxPercent / avgPercent;
-    if (
-        maxPercent < thresholds.maxPercent ||
-        avgPercent > thresholds.avgPercent ||
-        peakToAvg < thresholds.peakToAvgRatio
-    ) {
-        return undefined;
-    }
-    const max = Math.round(maxPercent);
-    const avg = Math.round(avgPercent);
-    const ratio = Math.round(peakToAvg * 10) / 10;
-    return {
-        id: 'AutoscaleCandidate',
-        rule: 'AutoscaleCandidate',
-        severity: 'Medium',
-        title: l10n.t('Workload looks like an autoscale candidate'),
-        rationale: clampRationale(
-            l10n.t(
-                'Over the last 7 days RU consumption peaked at {max}% but averaged just {avg}% — a {ratio}× peak-to-average burst on a mostly-idle baseline, while at least one container uses manual throughput. A tall peak over a low average is exactly the duty cycle where autoscale tracks demand more cheaply than fixed throughput sized for the peak.',
-                { max, avg, ratio },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Enable autoscale on the bursty containers so provisioned RU/s follows demand between the configured floor and ceiling.',
-        ),
-        thresholdReference: l10n.t('Threshold: peak ≥ {max}%, average ≤ {avg}%, and peak/average ≥ {ratio}×', {
-            max: thresholds.maxPercent,
-            avg: thresholds.avgPercent,
-            ratio: thresholds.peakToAvgRatio,
-        }),
-    };
-}
-
-const GIB = 1024 ** 3;
-const MS_PER_DAY = 24 * 60 * 60 * 1000;
-
-// Physical-partition storage split ceiling (50 GiB) — the wall StorageGrowthRisk projects to. Azure platform
-// limit, not a heuristic:
-// https://learn.microsoft.com/azure/cosmos-db/partitioning#physical-partitions
-const PARTITION_STORAGE_LIMIT_BYTES = 50 * GIB;
-/** Below this current size a partition is immaterial and ignored (a tiny partition near the wall is not a concern). */
-const STORAGE_GROWTH_MIN_MATERIAL_BYTES = 1 * GIB;
-/** Below this slope the growth is noise, not a trend — avoids projecting a wall from flat/jittery series. */
-const STORAGE_GROWTH_MIN_SLOPE_BYTES_PER_DAY = 0.1 * GIB;
-/** Days-to-limit at or below which StorageGrowthRisk ranks High. */
-const STORAGE_GROWTH_HIGH_DAYS = 30;
-/** Days-to-limit at or below which StorageGrowthRisk ranks Medium (above it, up to the horizon, Low). */
-const STORAGE_GROWTH_MEDIUM_DAYS = 90;
-
-/** Below this busiest-partition size, storage skew is immaterial (a balanced-but-tiny split is not a concern). */
-const STORAGE_SKEW_MIN_BUSIEST_BYTES = 1 * GIB;
-/** Busiest-partition size (≥ 80% of the 50 GiB ceiling) at or above which StorageSkewRisk ranks High. */
-const STORAGE_SKEW_HIGH_BYTES = 40 * GIB;
-/** Busiest-partition size (≥ 50% of the 50 GiB ceiling) at or above which StorageSkewRisk ranks Medium. */
-const STORAGE_SKEW_MEDIUM_BYTES = 25 * GIB;
-
-/** A physical partition's storage series (oldest → newest) for the growth/skew rules. */
-export type StoragePartitionSeries = PartitionStorageSeries;
-
-/** Per-container physical-partition storage series for the StorageGrowthRisk (and StorageSkewRisk) rules. */
-export interface ContainerStorageInput {
-    databaseId: string;
-    containerId: string;
-    partitions: readonly StoragePartitionSeries[];
-}
-
-/** Latest (newest-timestamp) size in bytes of a partition's series, or undefined when it has no samples. Pure. */
-function latestBytes(series: StoragePartitionSeries): number | undefined {
-    let latest: { timestamp: number; bytes: number } | undefined;
-    for (const sample of series.samples) {
-        if (!Number.isFinite(sample.bytes) || !Number.isFinite(sample.timestamp)) {
-            continue;
-        }
-        if (latest === undefined || sample.timestamp >= latest.timestamp) {
-            latest = sample;
-        }
-    }
-    return latest?.bytes;
-}
-
-/**
- * Least-squares slope of a partition's storage series in bytes/day. Returns
- * `undefined` when there are fewer than two datapoints at distinct times (no
- * trend can be fit). Fitting a trajectory — rather than a raw last-minus-first
- * delta — is robust to a single noisy endpoint. Pure.
- */
-export function storageGrowthSlopeBytesPerDay(samples: readonly PartitionStorageSample[]): number | undefined {
-    const clean = samples.filter((s) => Number.isFinite(s.timestamp) && Number.isFinite(s.bytes));
-    if (clean.length < 2) {
-        return undefined;
-    }
-    const xsDays = clean.map((s) => s.timestamp / MS_PER_DAY);
-    const ys = clean.map((s) => s.bytes);
-    const meanX = mean(xsDays);
-    const meanY = mean(ys);
-    let numerator = 0;
-    let denominator = 0;
-    for (let i = 0; i < clean.length; i++) {
-        const dx = xsDays[i] - meanX;
-        numerator += dx * (ys[i] - meanY);
-        denominator += dx * dx;
-    }
-    if (denominator === 0) {
-        return undefined;
-    }
-    return numerator / denominator;
-}
-
-/**
- * Projected days for a partition at `currentBytes` growing at `slopeBytesPerDay`
- * to reach `limitBytes`. Returns 0 if it is already at/over the limit, and
- * `undefined` for a flat or shrinking trajectory (it never reaches the wall).
- * Pure.
- */
-export function daysToStorageLimit(
-    currentBytes: number,
-    slopeBytesPerDay: number,
-    limitBytes: number,
-): number | undefined {
-    if (!Number.isFinite(slopeBytesPerDay) || slopeBytesPerDay <= 0) {
-        return undefined;
-    }
-    if (currentBytes >= limitBytes) {
-        return 0;
-    }
-    return (limitBytes - currentBytes) / slopeBytesPerDay;
-}
-
-/**
- * StorageGrowthRisk (best-effort): the soonest-to-fill physical partition in a
- * container is on a least-squares trajectory to the 50 GiB split ceiling within
- * the configured horizon. Unlike an absolute bytes-added trigger, this answers
- * "how soon do you hit the wall" — a 2 TB container adding 10 GiB/week spread over
- * dozens of partitions is decades away and does not fire, while a 40 GiB partition
- * growing steadily does. Immaterial (< 1 GiB) and flat/noisy (< 0.1 GiB/day)
- * partitions are ignored. Severity grades by horizon. Pure.
- */
-export function evaluateStorageGrowthRisk(
-    input: ContainerStorageInput,
-    horizonDays: number,
-): DerivedAdvisory | undefined {
-    if (horizonDays <= 0) {
-        return undefined;
-    }
-    let soonest: { days: number; currentBytes: number; slopeBytesPerDay: number } | undefined;
-    for (const partition of input.partitions) {
-        const currentBytes = latestBytes(partition);
-        if (currentBytes === undefined || currentBytes < STORAGE_GROWTH_MIN_MATERIAL_BYTES) {
-            continue;
-        }
-        const slopeBytesPerDay = storageGrowthSlopeBytesPerDay(partition.samples);
-        if (slopeBytesPerDay === undefined || slopeBytesPerDay < STORAGE_GROWTH_MIN_SLOPE_BYTES_PER_DAY) {
-            continue;
-        }
-        const days = daysToStorageLimit(currentBytes, slopeBytesPerDay, PARTITION_STORAGE_LIMIT_BYTES);
-        if (days === undefined) {
-            continue;
-        }
-        if (soonest === undefined || days < soonest.days) {
-            soonest = { days, currentBytes, slopeBytesPerDay };
-        }
-    }
-    if (soonest === undefined || soonest.days > horizonDays) {
-        return undefined;
-    }
-
-    const severity: DerivedAdvisorySeverity =
-        soonest.days <= STORAGE_GROWTH_HIGH_DAYS
-            ? 'High'
-            : soonest.days <= STORAGE_GROWTH_MEDIUM_DAYS
-              ? 'Medium'
-              : 'Low';
-    const scope = containerKey(input.databaseId, input.containerId);
-    const days = Math.max(0, Math.round(soonest.days));
-    const currentGiB = Math.round((soonest.currentBytes / GIB) * 10) / 10;
-    const perDayGiB = Math.round((soonest.slopeBytesPerDay / GIB) * 100) / 100;
-    const horizon = Math.round(horizonDays);
-    return {
-        id: `StorageGrowthRisk:${scope}`,
-        rule: 'StorageGrowthRisk',
-        severity,
-        title: l10n.t('Physical partition approaching its storage limit in {container}', {
-            container: input.containerId,
-        }),
-        rationale: clampRationale(
-            l10n.t(
-                'The fastest-growing physical partition in "{container}" is at {current} GiB and growing about {perDay} GiB/day — on that trajectory it reaches the 50 GiB physical-partition split ceiling in roughly {days} days, within the {horizon}-day risk horizon. A partition that hits the wall can throttle or block writes until it splits.',
-                { container: input.containerId, current: currentGiB, perDay: perDayGiB, days, horizon },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Confirm the partition key spreads new data across partitions, archive or delete cold data, and ensure large logical keys stay well under the 20 GiB per-key cap so partitions split cleanly.',
-        ),
-        thresholdReference: l10n.t('Threshold: projected < {horizon} days to the 50 GiB partition ceiling', {
-            horizon,
-        }),
-        scope,
-    };
-}
-
-/**
- * Balance ratio (`min ÷ max`) of a set of physical-partition sizes: 1.0 is
- * perfectly balanced, values near 0 mean one partition dwarfs its siblings.
- * Returns `undefined` for fewer than two sizes or a non-positive max (nothing to
- * compare). Pure.
- */
-export function balanceRatio(sizesBytes: readonly number[]): number | undefined {
-    const clean = sizesBytes.filter((v) => Number.isFinite(v) && v >= 0);
-    if (clean.length < 2) {
-        return undefined;
-    }
-    const max = Math.max(...clean);
-    const min = Math.min(...clean);
-    if (max <= 0) {
-        return undefined;
-    }
-    return min / max;
-}
-
-/**
- * StorageSkewRisk (best-effort): physical-partition sizes are imbalanced
- * (`min/max` below the configured balance ratio) *and* the busiest partition is
- * material (≥ 1 GiB). Unlike a raw storage-share cutoff — which is
- * partition-count-dependent and size-blind (a balanced 40 MiB/60 MiB two-way
- * split trips a 35% share) — the balance ratio only fires on genuine imbalance,
- * and severity grades by the busiest partition's proximity to the 50 GiB split
- * ceiling (≥ 80% → High, ≥ 50% → Medium). Balanced-but-large partitions are
- * healthy: they simply split as they grow. Pure.
- */
-export function evaluateStorageSkewRisk(
-    input: ContainerStorageInput,
-    balanceThreshold: number,
-): DerivedAdvisory | undefined {
-    if (balanceThreshold <= 0) {
-        return undefined;
-    }
-    const sizes = input.partitions
-        .map((partition) => latestBytes(partition))
-        .filter((bytes): bytes is number => bytes !== undefined);
-    const ratio = balanceRatio(sizes);
-    if (ratio === undefined || ratio >= balanceThreshold) {
-        return undefined;
-    }
-    const busiestBytes = Math.max(...sizes);
-    if (busiestBytes < STORAGE_SKEW_MIN_BUSIEST_BYTES) {
-        return undefined;
-    }
-
-    const severity: DerivedAdvisorySeverity =
-        busiestBytes >= STORAGE_SKEW_HIGH_BYTES ? 'High' : busiestBytes >= STORAGE_SKEW_MEDIUM_BYTES ? 'Medium' : 'Low';
-    const scope = containerKey(input.databaseId, input.containerId);
-    const busiestGiB = Math.round((busiestBytes / GIB) * 10) / 10;
-    const balance = Math.round(ratio * 100) / 100;
-    const threshold = Math.round(balanceThreshold * 100) / 100;
-    const ceilingPercent = Math.round((busiestBytes / PARTITION_STORAGE_LIMIT_BYTES) * 100);
-    return {
-        id: `StorageSkewRisk:${scope}`,
-        rule: 'StorageSkewRisk',
-        severity,
-        title: l10n.t('Uneven physical-partition storage in {container}', { container: input.containerId }),
-        rationale: clampRationale(
-            l10n.t(
-                'Physical-partition storage in "{container}" is uneven — the coolest partition holds only {balance}× the busiest (below the {threshold}× balance floor), and the busiest is at {busiest} GiB, about {ceiling}% of the 50 GiB split ceiling. A single oversized partition hits the split wall long before its balanced siblings, capping the container.',
-                { container: input.containerId, balance, threshold, busiest: busiestGiB, ceiling: ceilingPercent },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Review the partition key for a higher-cardinality choice so data spreads evenly, and split or archive the oversized logical keys concentrating storage on one partition.',
-        ),
-        thresholdReference: l10n.t('Threshold: min/max partition size < {threshold}× with busiest ≥ 1 GiB', {
-            threshold,
-        }),
-        scope,
-    };
-}
-
-/** Per-container storage inputs for the IndexingCostRisk rule. */
-export interface IndexingUsageInput {
-    databaseId: string;
-    containerId: string;
-    /** Latest `IndexUsage` bytes for the container, when reported. */
-    indexUsageBytes?: number;
-    /** Latest `DataUsage` bytes for the container, when reported. */
-    dataUsageBytes?: number;
-    /** Number of paths excluded from indexing (0 ⇒ indexes everything). */
-    excludedPathCount: number;
-}
-
-/**
- * IndexingCostRisk (best-effort): index storage is a large fraction of data
- * storage AND the container excludes no paths (indexes everything). Pure.
- */
-export function evaluateIndexingCostRisk(
-    input: IndexingUsageInput,
-    ratioThreshold: number,
-): DerivedAdvisory | undefined {
-    const { indexUsageBytes, dataUsageBytes, excludedPathCount } = input;
-    if (
-        excludedPathCount > 0 ||
-        dataUsageBytes === undefined ||
-        dataUsageBytes <= 0 ||
-        indexUsageBytes === undefined ||
-        indexUsageBytes < 0
-    ) {
-        return undefined;
-    }
-    const ratio = indexUsageBytes / dataUsageBytes;
-    if (ratio <= ratioThreshold) {
-        return undefined;
-    }
-    const scope = containerKey(input.databaseId, input.containerId);
-    const percent = Math.round(ratio * 100);
-    const threshold = Math.round(ratioThreshold * 100);
-    return {
-        id: `IndexingCostRisk:${scope}`,
-        rule: 'IndexingCostRisk',
-        severity: 'Low',
-        title: l10n.t('High indexing overhead in {container}', { container: input.containerId }),
-        rationale: clampRationale(
-            l10n.t(
-                'Index storage in "{container}" is {percent}% of its data size, above the {threshold}% threshold, and the container indexes every path (no exclusions). Indexing every property inflates storage and the RU cost of every write.',
-                { container: input.containerId, percent, threshold },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Exclude paths that are never used in filters from the indexing policy to cut write RU charges and index storage.',
-        ),
-        thresholdReference: l10n.t('Threshold: index/data storage > {threshold}%', { threshold }),
-        scope,
-    };
-}
-
-// ─── Account-scoped configuration advisories (ARM config only) ─────────────────────
-//
-// DX-016 and DX-008 read only the account's ARM configuration — no workload, no metrics — and surface
-// qualitative antipatterns (they explain the consequence and ask the customer to confirm intent; they assert
-// no RU/s or dollar figure). Both are account-scoped, so at most one of each fires per account.
-
-/** Consistency levels stronger than the Session default, normalised to lowercase (no separators). */
-const PREMIUM_CONSISTENCY_LEVELS: ReadonlySet<string> = new Set(['strong', 'boundedstaleness']);
-
-/** Cosmos DB APIs whose drivers cannot use multi-region-write failover, so the flag is always a misconfiguration. */
-const NON_MULTI_WRITE_APIS: ReadonlySet<string> = new Set(['mongo', 'cassandra', 'gremlin', 'table']);
-
-const API_LABEL: Record<string, string> = {
-    mongo: 'MongoDB',
-    cassandra: 'Cassandra',
-    gremlin: 'Gremlin',
-    table: 'Table',
-};
-
-/** Already-known ARM account configuration the config-only advisories (DX-016 / DX-008) read. */
-export interface AccountConfigInput {
-    /** Account name, for environment inference (DX-008). */
-    accountName: string;
-    /** Resource tags, for environment inference (DX-008). */
-    tags?: Record<string, string | undefined>;
-    /** Subscription display name, for environment inference (DX-008). */
-    subscriptionName?: string;
-    /** `defaultConsistencyLevel` from the ARM consistency policy (e.g. `Session`, `Strong`). */
-    consistencyLevel?: string;
-    /** Number of regions enabled for the account. */
-    regionCount: number;
-    /** `enableMultipleWriteLocations` from ARM. */
-    multiRegionWritesEnabled: boolean;
-    /** Number of write regions the account has. */
-    writeRegionCount: number;
-    /**
-     * Canonical API family (`core`, `mongo`, `cassandra`, `gremlin`, `table`). The derived-advisory engine only
-     * runs for `core` accounts, so the wrong-API DX-008 branch is effectively dead here, but it is kept faithful
-     * to CODA so the rule is correct if the engine is ever opened to other APIs.
-     */
-    apiKind: string;
-}
-
-/** Normalises an ARM consistency level (any casing / separators) to the lowercase token used for comparison. */
-function normalizeConsistency(level: string | undefined): string {
-    return (level ?? '').split('.').pop()!.replace(/[_\s]/g, '').trim().toLowerCase();
-}
-
-/**
- * ExpensiveConsistency (CODA DX-016): the account runs a premium consistency level (Strong or Bounded Staleness)
- * across two or more regions. Across multiple regions these levels make writes wait on additional regions, raising
- * write latency and tightening availability; the default Session consistency suits most workloads. On a single
- * region the cross-region cost does not apply, so it never fires there. Strong is Medium, Bounded Staleness is Low.
- * Pure.
- */
-export function evaluateExpensiveConsistency(config: AccountConfigInput): DerivedAdvisory | undefined {
-    const level = normalizeConsistency(config.consistencyLevel);
-    if (!PREMIUM_CONSISTENCY_LEVELS.has(level) || config.regionCount < 2) {
-        return undefined;
-    }
-    const isStrong = level === 'strong';
-    const displayLevel = config.consistencyLevel ?? (isStrong ? 'Strong' : 'Bounded Staleness');
-    return {
-        id: 'ExpensiveConsistency',
-        rule: 'ExpensiveConsistency',
-        severity: isStrong ? 'Medium' : 'Low',
-        title: l10n.t('Premium consistency across regions'),
-        rationale: clampRationale(
-            l10n.t(
-                'This account uses {level} consistency across {regions} regions. Across multiple regions this level raises write latency and tightens availability (Strong makes writes wait on additional regions), and the default Session consistency is sufficient for most workloads.',
-                { level: displayLevel, regions: config.regionCount },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Confirm the application genuinely needs this consistency level; if not, relax it to Session (or Bounded Staleness) after validating the application’s tolerance.',
-        ),
-        thresholdReference: l10n.t('Threshold: Strong or Bounded Staleness across ≥ 2 regions'),
-    };
-}
-
-/**
- * MultiRegionWriteAntipattern (CODA DX-008): multi-region writes are enabled where the RTO=0 write-failover
- * benefit cannot or need not apply. The wrong-API case (Mongo/Cassandra/Gremlin/Table drivers cannot use the
- * failover) is highest confidence; a single write region with the flag set is a latent tripwire (no benefit today,
- * doubled write cost the moment a second write region is added); non-production with ≥ 2 write regions does not
- * need RTO=0 write failover. Two-or-more write regions in production is legitimate active-active HA and never
- * fires. Pure.
- */
-export function evaluateMultiRegionWrites(config: AccountConfigInput): DerivedAdvisory | undefined {
-    if (!config.multiRegionWritesEnabled) {
-        return undefined;
-    }
-
-    const base = { id: 'MultiRegionWriteAntipattern', rule: 'MultiRegionWriteAntipattern' as const };
-    const disableAction = l10n.t('Disable multi-region writes unless the write-failover benefit genuinely applies.');
-
-    // Wrong-API: the failover benefit is impossible regardless of environment — highest confidence.
-    if (NON_MULTI_WRITE_APIS.has(config.apiKind)) {
-        const apiLabel = API_LABEL[config.apiKind];
-        return {
-            ...base,
-            severity: 'High',
-            title: l10n.t('Multi-region writes on a {api} account', { api: apiLabel }),
-            rationale: clampRationale(
-                l10n.t(
-                    'Multi-region writes are enabled on a {api} API account, but the {api} drivers cannot use Cosmos DB’s multi-region-write failover — so the RTO=0 HA benefit never materialises while the account still carries the write-conflict surface (and 2× write cost across any second write region).',
-                    { api: apiLabel },
-                ),
-            ),
-            suggestedAction: disableAction,
-            thresholdReference: l10n.t('Threshold: multi-region writes enabled on a non-SQL API account'),
-        };
-    }
-
-    // Single write region + the flag set = latent tripwire regardless of environment: no benefit today (write
-    // failover needs a second region), and write cost doubles the moment a second write region is added.
-    if (config.writeRegionCount < 2) {
-        return {
-            ...base,
-            severity: 'Low',
-            title: l10n.t('Multi-region writes with a single write region'),
-            rationale: clampRationale(
-                l10n.t(
-                    'Multi-region writes are enabled but the account has a single write region, so the feature delivers no benefit today (RTO=0 write failover needs a second region), while write cost doubles the moment a second write region is added.',
-                ),
-            ),
-            suggestedAction: l10n.t(
-                'Disable multi-region writes to remove the tripwire unless a second write region is imminent and justified.',
-            ),
-            thresholdReference: l10n.t('Threshold: multi-region writes enabled with a single write region'),
-        };
-    }
-
-    // Two or more write regions: legitimate active-active HA in production; only flagged in non-production.
-    const env = classifyEnvironment(config.accountName, config.tags, config.subscriptionName);
-    if (!env.isNonProd) {
-        return undefined;
-    }
-    return {
-        ...base,
-        severity: 'Medium',
-        title: l10n.t('Multi-region writes on a non-production account'),
-        rationale: clampRationale(
-            l10n.t(
-                'Multi-region writes are enabled across {regions} write regions on a non-production account, which does not need RTO=0 write failover. This doubles write cost and adds write-conflict surface for no benefit.',
-                { regions: config.writeRegionCount },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Disable multi-region writes (and consider dropping the second write region) on dev/test/stage accounts.',
-        ),
-        thresholdReference: l10n.t('Threshold: multi-region writes across ≥ 2 write regions on a non-prod account'),
-    };
-}
-
-// DX-004/011: autoscale bills between 10% and 100% of the configured max (`0.1*Tmax ≤ T ≤ Tmax`), so an idle
-// or over-sized autoscale container only recovers its idle floor, never the full max:
-// https://learn.microsoft.com/azure/cosmos-db/provision-throughput-autoscale#how-autoscale-throughput-works
-const AUTOSCALE_IDLE_FRACTION = 0.1;
-
-/** Already-fetched idle signal for one container over the 30-day window (DX-004). */
-export interface IdleContainerInput {
-    databaseId: string;
-    containerId: string;
-    /** Peak RU consumed in any single bucket (max of `TotalRequestUnits`, Total). 0 for an all-zero/empty series. */
-    peakRuPerBucket: number;
-    /** Throughput billing mode; only manual (dedicated/shared) and autoscale offers are recoverable. */
-    throughputMode: ThroughputMode;
-    /** The container's configured provisioned RU/s (autoscale max for autoscale). */
-    provisionedRu?: number;
-}
-
-/**
- * IdleContainer (CODA DX-004): a container that is provisioned but serves ~no traffic — its whole reservation is
- * pure 24/7 waste. The signal is the **peak RU consumed in any single bucket** (`max(TotalRequestUnits, Total)`)
- * over a long 30-day window: a container that never spends more than a tiny system floor in *any* bucket did no
- * real work, so it is idle. This is deliberately spike-aware — a container silent for 29 days that runs one real
- * batch on day 30 shows a large peak and is correctly not flagged. Unlike the utilisation rules an empty/all-zero
- * series is the *positive* idle signal (Azure Monitor pads idle windows with zeros), not an abstention.
- * Mode-aware recoverable capacity: manual idle bills the full provisioned RU/s (all recoverable); autoscale idle
- * scales to ~10% of its max (that floor is what decommission recovers). Severity is capacity materiality. Pure.
- */
-export function evaluateIdleContainer(
-    input: IdleContainerInput,
-    maxIdlePeakRu: number,
-    scopeProvisionedRu?: number,
-): DerivedAdvisory | undefined {
-    const isAutoscale = input.throughputMode === 'autoscale';
-    const isManual = MANUAL_THROUGHPUT_MODES.has(input.throughputMode);
-    // Only manual/autoscale containers have a provisioned offer to recover; serverless has none.
-    if ((!isAutoscale && !isManual) || input.provisionedRu === undefined || input.provisionedRu <= 0) {
-        return undefined;
-    }
-    // Consumed real RU in some bucket ⇒ active, not idle — DX-001 right-sizes it instead.
-    if (input.peakRuPerBucket > maxIdlePeakRu) {
-        return undefined;
-    }
-    const recoverableRu = isAutoscale ? AUTOSCALE_IDLE_FRACTION * input.provisionedRu : input.provisionedRu;
-    const severity = capacityMaterialitySeverity(recoverableRu, scopeProvisionedRu);
-    const scope = containerKey(input.databaseId, input.containerId);
-    const peak = Math.round(input.peakRuPerBucket);
-    const recoverable = Math.round(recoverableRu);
-    const provisioned = Math.round(input.provisionedRu);
-    const rationale = isAutoscale
-        ? l10n.t(
-              'Container "{container}" consumed at most {peak} RU in any single bucket over the last 30 days, at or below the {threshold} RU idle floor — it serves ~no traffic. Decommissioning it recovers the autoscale idle floor (~{recoverable} RU/s of its {provisioned} RU/s max).',
-              { container: input.containerId, peak, threshold: Math.round(maxIdlePeakRu), recoverable, provisioned },
-          )
-        : l10n.t(
-              'Container "{container}" consumed at most {peak} RU in any single bucket over the last 30 days, at or below the {threshold} RU idle floor — it serves ~no traffic. Its full {recoverable} RU/s provisioned capacity is recoverable.',
-              { container: input.containerId, peak, threshold: Math.round(maxIdlePeakRu), recoverable },
-          );
-    return {
-        id: `IdleContainer:${scope}`,
-        rule: 'IdleContainer',
-        severity,
-        title: l10n.t('Container looks idle'),
-        rationale: clampRationale(rationale),
-        suggestedAction: l10n.t(
-            'Review "{container}" for decommission, a TTL policy, or archival; confirm it is genuinely unused before removing it.',
-            { container: input.containerId },
-        ),
-        thresholdReference: l10n.t('Threshold: peak ≤ {threshold} RU per bucket over 30 days', {
-            threshold: Math.round(maxIdlePeakRu),
-        }),
-        scope,
-    };
-}
-
-// DX-009: each physical partition serves up to 10,000 RU/s and stores up to 50 GiB. Both are Azure platform
-// limits, so the needed-partition math below is grounded rather than a heuristic:
-// https://learn.microsoft.com/azure/cosmos-db/partitioning#physical-partitions
-const RU_PER_PARTITION = 10_000;
-const BYTES_PER_PARTITION = 50 * GIB;
-
-/** Already-fetched partition-count + storage signal for one container (DX-009). */
-export interface PartitionMergeInput {
-    databaseId: string;
-    containerId: string;
-    /** Actual physical partition count (max of `PhysicalPartitionCount`). 0 ⇒ no reading. */
-    actualPartitions: number;
-    /** The container's configured provisioned RU/s (autoscale max or manual). */
-    provisionedRu?: number;
-    /** Current data storage in bytes (max of `DataUsage`). */
-    dataUsageBytes?: number;
-}
-
-/** The smallest physical-partition count that satisfies both the RU/s and storage needs (CODA DX-009). Pure. */
-function neededPartitions(provisionedRu: number, storageBytes: number): number {
-    const forRu = provisionedRu > 0 ? Math.ceil(provisionedRu / RU_PER_PARTITION) : 1;
-    const forStorage = storageBytes > 0 ? Math.ceil(storageBytes / BYTES_PER_PARTITION) : 1;
-    return Math.max(forRu, forStorage, 1);
-}
-
-/**
- * PartitionMergeCandidate (CODA DX-009): a container has **more physical partitions than its throughput and
- * storage need** — typically a leftover from once having been scaled to a high RU/s (or large) and then scaled
- * back down. Each partition then gets a thin slice of RU/s, so a single hot partition can throttle while the
- * container as a whole looks under-utilised. `needed = max(ceil(RU / 10,000), ceil(bytes / 50 GiB), 1)`; the
- * container is over-partitioned when `actual > needed` and `actual ≥ 2` (a single partition is never flagged).
- * Severity is Medium when actual ≥ 2× needed (the slice is very thin), else Low. Merge is a platform operation.
- * Pure.
- */
-export function evaluatePartitionMergeCandidate(input: PartitionMergeInput): DerivedAdvisory | undefined {
-    const actual = input.actualPartitions;
-    // 0 ⇒ no partition reading (insufficient data); 1 ⇒ single partition can never be merged.
-    if (actual <= 1) {
-        return undefined;
-    }
-    const provisionedRu = input.provisionedRu ?? 0;
-    const storageBytes = input.dataUsageBytes ?? 0;
-    const needed = neededPartitions(provisionedRu, storageBytes);
-    if (actual <= needed) {
-        return undefined;
-    }
-    const severity: DerivedAdvisorySeverity = actual >= 2 * needed ? 'Medium' : 'Low';
-    const scope = containerKey(input.databaseId, input.containerId);
-    const ruSlice = provisionedRu > 0 ? Math.round(provisionedRu / actual) : 0;
-    const gib = Math.round((storageBytes / GIB) * 10) / 10;
-    return {
-        id: `PartitionMergeCandidate:${scope}`,
-        rule: 'PartitionMergeCandidate',
-        severity,
-        title: l10n.t('Container may be over-partitioned'),
-        rationale: clampRationale(
-            l10n.t(
-                'Container "{container}" has {actual} physical partitions but its {ru} RU/s and {gib} GiB only need {needed}. Each partition therefore gets just ~{slice} RU/s, so a single hot partition can throttle while the container looks under-utilised. This usually follows scaling RU/s down (or deleting data) after a high-throughput or large period.',
-                { container: input.containerId, actual, ru: Math.round(provisionedRu), gib, needed, slice: ruSlice },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Merge the physical partitions (Azure Portal or CLI) to consolidate the RU/s back onto fewer partitions.',
-        ),
-        thresholdReference: l10n.t('Threshold: physical partitions > max(ceil(RU/10,000), ceil(storage/50 GiB), 1)'),
-        scope,
-    };
-}
-
-/** Already-fetched autoscale duty-cycle signal for one container (DX-011 / DX-013). */
-export interface AutoscaleUtilizationInput {
-    databaseId: string;
-    containerId: string;
-    /** Peak of the `AutoscaledRU`-as-%-of-configured-max series over the window (0..100). */
-    peakPercent: number;
-    /** Average of the same series (0..100), for the duty-cycle read. */
-    avgPercent: number;
-    /** Number of samples in the series (0 ⇒ no data). */
-    sampleCount: number;
-    /** The container's configured autoscale max RU/s. */
-    configuredMaxRu?: number;
-}
-
-/**
- * AutoscaleMaxOverProvisioned (CODA DX-011): for autoscale, `NormalizedRUConsumption` measures against the
- * *currently-scaled* value, so it cannot tell whether the configured **max** is too high. The correct signal is
- * `AutoscaledRU` (what autoscale actually provisioned) as a % of the configured max. We band on the **peak** of
- * that series (autoscale's max must cover legitimate spikes, so lowering it below a real peak would throttle):
- * the max is over-provisioned when the busiest sample stayed below the moderate band. Autoscale bills the idle
- * floor (10% of max) even at zero load, so lowering the max only recovers the floor reduction, not the full
- * headroom. Severity is capacity materiality on that recoverable floor. Pure.
- */
-export function evaluateAutoscaleMaxOverProvisioned(
-    input: AutoscaleUtilizationInput,
-    bandPercent: number,
-    scopeProvisionedRu?: number,
-): DerivedAdvisory | undefined {
-    if (input.sampleCount === 0 || input.configuredMaxRu === undefined || input.configuredMaxRu <= 0) {
-        return undefined;
-    }
-    const peak = input.peakPercent;
-    // Peak at/above the saturation guard, or at/above the moderate band, means the max is genuinely needed.
-    if (peak >= OVERPROVISIONING_PEAK_GUARD_PCT || peak >= bandPercent) {
-        return undefined;
-    }
-    const max = input.configuredMaxRu;
-    const peakConsumed = (peak / 100) * max;
-    const recommended = Math.min(
-        max,
-        Math.max(OVERPROVISIONING_MIN_RU, Math.ceil(peakConsumed * OVERPROVISIONING_HEADROOM), Math.ceil(peakConsumed)),
-    );
-    const floor = (ru: number): number => Math.max(OVERPROVISIONING_MIN_RU, AUTOSCALE_IDLE_FRACTION * ru);
-    const wastedRu = Math.max(0, Math.round(floor(max) - floor(recommended)));
-    const severity = capacityMaterialitySeverity(wastedRu, scopeProvisionedRu);
-    const scope = containerKey(input.databaseId, input.containerId);
-    const peakPct = Math.round(peak);
-    const recommend = Math.round(recommended);
-    return {
-        id: `AutoscaleMaxOverProvisioned:${scope}`,
-        rule: 'AutoscaleMaxOverProvisioned',
-        severity,
-        title: l10n.t('Autoscale max may be set too high'),
-        rationale: clampRationale(
-            l10n.t(
-                'Autoscale on "{container}" never provisioned more than {peak}% of its {max} RU/s max over the window, below the {band}% band. Autoscale bills a 10% idle floor of the max even at zero load, so an over-high max wastes ~{wasted} RU/s of that floor.',
-                {
-                    container: input.containerId,
-                    peak: peakPct,
-                    max: Math.round(max),
-                    band: Math.round(bandPercent),
-                    wasted: wastedRu,
-                },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Lower the autoscale max toward the peak-provisioned level (≈ {recommended} RU/s) so the idle floor tracks real demand while still covering the peak.',
-            { recommended: recommend },
-        ),
-        thresholdReference: l10n.t('Threshold: peak AutoscaledRU < {band}% of the configured max', {
-            band: Math.round(bandPercent),
-        }),
-        scope,
-    };
-}
-
-/**
- * AutoscaleToManualCandidate (CODA DX-013): an *autoscale* container whose load is **steady-high** — average
- * sustained near the max with little spikiness. Autoscale carries a per-RU premium over manual, so a workload
- * that stays high does not benefit from scaling and would be cheaper on manual at a fixed RU/s. Steady ⇔ average
- * ≥ `avgFloorPercent` of the max AND peak-to-average ≤ `peakToAvgCeiling` (flat, not bursty). Qualitative (a
- * pricing-mode change, not a capacity recovery), so severity is Medium. Pure.
- */
-export function evaluateAutoscaleToManualCandidate(
-    input: AutoscaleUtilizationInput,
-    avgFloorPercent: number,
-    peakToAvgCeiling: number,
-): DerivedAdvisory | undefined {
-    if (input.sampleCount === 0 || input.avgPercent <= 0) {
-        return undefined;
-    }
-    const peakToAvg = input.peakPercent / input.avgPercent;
-    if (input.avgPercent < avgFloorPercent || peakToAvg > peakToAvgCeiling) {
-        return undefined;
-    }
-    const scope = containerKey(input.databaseId, input.containerId);
-    const avg = Math.round(input.avgPercent);
-    const peak = Math.round(input.peakPercent);
-    const ratio = Math.round(peakToAvg * 10) / 10;
-    return {
-        id: `AutoscaleToManualCandidate:${scope}`,
-        rule: 'AutoscaleToManualCandidate',
-        severity: 'Medium',
-        title: l10n.t('Autoscale container looks steady — manual may be cheaper'),
-        rationale: clampRationale(
-            l10n.t(
-                'Autoscale on "{container}" averaged {avg}% of its max with a peak of {peak}% — a {ratio}× peak-to-average, so the load is steady and high with little spikiness. Autoscale carries a per-RU premium over manual, and a workload that stays high does not benefit from scaling.',
-                { container: input.containerId, avg, peak, ratio },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Consider converting "{container}" to manual throughput at a fixed RU/s near the sustained level to drop the autoscale premium.',
-            { container: input.containerId },
-        ),
-        thresholdReference: l10n.t('Threshold: average ≥ {avg}% of max and peak/average ≤ {ratio}×', {
-            avg: Math.round(avgFloorPercent),
-            ratio: Math.round(peakToAvgCeiling * 10) / 10,
-        }),
-        scope,
-    };
-}
-
-/** DX-014: minimum number of buckets needed to judge the account-total RU shape. */
-const SERVERLESS_MIN_POINTS = 24;
-
-/** Already-computed account-total consumed-RU shape over the 30-day window (DX-014). */
-export interface ServerlessCandidateInput {
-    /** Account-total average consumed RU/s over the window. */
-    avgRuPerSec: number;
-    /** Account-total peak consumed RU/s over the window. */
-    peakRuPerSec: number;
-    /** Number of buckets sampled. */
-    sampleCount: number;
-    /** True when the account already uses serverless capacity mode (guard: never a candidate). */
-    isServerless: boolean;
-}
-
-/** Tunable cutoffs for {@link evaluateServerlessCandidate}. */
-export interface ServerlessCandidateThresholds {
-    /** Average-to-peak RU/s ratio below which the workload is sporadic enough for serverless. */
-    sporadicRatio: number;
-    /** Account-total peak (RU/s) above which the workload is real enough to evaluate serverless (else it is idle). */
-    peakFloorRuPerSec: number;
-    /** Account-total peak (RU/s) at or below which serverless still fits (single-partition ceiling). */
-    peakCeilingRuPerSec: number;
-}
-
-/**
- * ServerlessCandidate (CODA DX-014): the account's total consumed throughput is **low and sporadic** — a small
- * peak with long idle stretches (a low average-to-peak ratio). Serverless bills per RU with no provisioned floor,
- * so it wins for this shape. Candidate ⇔ enough history AND peak in `(peakFloor, peakCeiling]` RU/s AND
- * average/peak < `sporadicRatio`. An account already on serverless, or one whose peak is below the floor (idle —
- * that is DX-004) or above the ceiling (too large for serverless), is not a candidate. Migration-class, so
- * qualitative Low severity. Account-scoped. Pure.
- */
-export function evaluateServerlessCandidate(
-    input: ServerlessCandidateInput,
-    thresholds: ServerlessCandidateThresholds,
-): DerivedAdvisory | undefined {
-    if (input.isServerless || input.sampleCount < SERVERLESS_MIN_POINTS) {
-        return undefined;
-    }
-    if (input.peakRuPerSec < thresholds.peakFloorRuPerSec || input.peakRuPerSec > thresholds.peakCeilingRuPerSec) {
-        return undefined;
-    }
-    const ratio = input.peakRuPerSec > 0 ? input.avgRuPerSec / input.peakRuPerSec : 0;
-    if (ratio >= thresholds.sporadicRatio) {
-        return undefined;
-    }
-    const peak = Math.round(input.peakRuPerSec);
-    const avg = Math.round(input.avgRuPerSec);
-    const ratioPct = Math.round(ratio * 100);
-    return {
-        id: 'ServerlessCandidate',
-        rule: 'ServerlessCandidate',
-        severity: 'Low',
-        title: l10n.t('Account may suit serverless'),
-        rationale: clampRationale(
-            l10n.t(
-                'Account throughput is low and sporadic — peak {peak} RU/s, average {avg} RU/s (a {ratio}% average-to-peak with long idle stretches). Serverless (pay-per-RU, no provisioned floor) may cost less for this shape.',
-                { peak, avg, ratio: ratioPct },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Evaluate serverless: it is migration-class (a new, single-region account, no shared-throughput databases, 5,000 RU/s per-partition cap), so validate those constraints fit before migrating data.',
-        ),
-        thresholdReference: l10n.t('Threshold: peak in ({floor}, {ceiling}] RU/s and average/peak < {ratio}', {
-            floor: Math.round(thresholds.peakFloorRuPerSec),
-            ceiling: Math.round(thresholds.peakCeilingRuPerSec),
-            ratio: thresholds.sporadicRatio,
-        }),
-    };
-}
-
-// ─── Tier-2 (Log Analytics) evaluators ───────────────────────────────────────────
-//
-// DX-002/003/007/010 are deterministic arithmetic over rows the `CDB*` Kusto tables already contain
-// (query fan-out, write-RU share, per-collection 429/RU disparity). The fetch-and-shape lives in
-// `derivedAdvisoryLogs.ts`; these functions are pure so they can be unit-tested against the same
-// synthetic shapes CODA's own offline detector tests use.
-
-/** One query shape (grouped by anonymized `QueryText`) and its fan-out over the window (CODA DX-002 `QueryShape`). */
-export interface QueryShapeInput {
-    /** Anonymized query text (identifiers → `p1`/`p2`, literals → `str1`), used only as display context. */
-    text: string;
-    /** Number of logical query executions of this shape. */
-    executions: number;
-    /** Average distinct physical partitions touched per execution (`avg(dcount(PartitionKeyRangeId))`). */
-    avgFanout: number;
-    /** Widest fan-out observed for this shape, the container's estimated physical-partition count. */
-    maxFanout: number;
-}
-
-/** Per-container cross-partition / shard-key input (DX-002 / DX-007). */
-export interface CrossPartitionInput {
-    databaseId: string;
-    containerId: string;
-    shapes: readonly QueryShapeInput[];
-}
-
-export interface CrossPartitionThresholds {
-    minQueries: number;
-    fanoutThreshold: number;
-    highPct: number;
-    medPct: number;
-}
-
-/** Folds the per-shape fan-out into total executions, cross-partition share, and the worst (most frequent) shape. Pure. */
-interface CrossPartitionStats {
-    totalQueries: number;
-    containerPartitions: number;
-    crossPartitionPct: number;
-    worst: QueryShapeInput | undefined;
-}
-
-function crossPartitionStats(shapes: readonly QueryShapeInput[], fanoutThreshold: number): CrossPartitionStats {
-    const totalQueries = shapes.reduce((sum, s) => sum + s.executions, 0);
-    const containerPartitions = shapes.reduce((max, s) => Math.max(max, s.maxFanout), 0);
-    const crossShapes = shapes.filter((s) => s.avgFanout >= fanoutThreshold);
-    const crossExec = crossShapes.reduce((sum, s) => sum + s.executions, 0);
-    const crossPartitionPct = totalQueries > 0 ? (100 * crossExec) / totalQueries : 0;
-    const worst = crossShapes.reduce<QueryShapeInput | undefined>(
-        (best, s) => (best === undefined || s.executions > best.executions ? s : best),
-        undefined,
-    );
-    return { totalQueries, containerPartitions, crossPartitionPct, worst };
-}
-
-/** Collapses whitespace and truncates an anonymized query shape for display. Pure. */
-function shortQueryShape(text: string): string {
-    const collapsed = text.split(/\s+/).filter(Boolean).join(' ');
-    return collapsed.length > 200 ? collapsed.slice(0, 200) + '…' : collapsed;
-}
-
-/**
- * CrossPartitionQuery (CODA DX-002): query shapes that do not filter on the container's partition key fan out
- * across every physical partition, multiplying RU and latency. Fires when the container has ≥ 2 physical
- * partitions, there is enough query traffic (`minQueries`), and the share of executions whose shape fans out
- * (average ≥ `fanoutThreshold` partitions) is at or above `medPct` — High at/above `highPct`, else Medium. The
- * surfaced text is the service-anonymized shape, so remediation is structural (add the partition key to the
- * filter). Container-scoped. Pure.
- */
-export function evaluateCrossPartitionQuery(
-    input: CrossPartitionInput,
-    thresholds: CrossPartitionThresholds,
-): DerivedAdvisory | undefined {
-    const { totalQueries, containerPartitions, crossPartitionPct, worst } = crossPartitionStats(
-        input.shapes,
-        thresholds.fanoutThreshold,
-    );
-    if (totalQueries < thresholds.minQueries || containerPartitions <= 1) {
-        return undefined;
-    }
-    if (crossPartitionPct < thresholds.medPct || worst === undefined) {
-        return undefined;
-    }
-    const severity: DerivedAdvisorySeverity = crossPartitionPct >= thresholds.highPct ? 'High' : 'Medium';
-    const pct = Math.round(crossPartitionPct);
-    const shape = shortQueryShape(worst.text);
-    return {
-        id: `CrossPartitionQuery/${containerKey(input.databaseId, input.containerId)}`,
-        rule: 'CrossPartitionQuery',
-        severity,
-        scope: containerKey(input.databaseId, input.containerId),
-        title: l10n.t('Cross-partition query fan-out in {container}', { container: input.containerId }),
-        rationale: clampRationale(
-            l10n.t(
-                '{pct}% of query executions fan out across partitions (the container has ~{partitions} physical partitions). The most frequent offender averages {fanout} partitions per run. Anonymized shape: {shape}',
-                { pct, partitions: containerPartitions, fanout: worst.avgFanout.toFixed(1), shape },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Add the container’s partition key to the query filter (or a composite index) so it targets a single logical partition. For access patterns that inherently filter on a different key, serve them from a secondary index or a change-feed-synced copy. Cross-partition fan-out multiplies RU and latency and does not scale as the container grows.',
-        ),
-        thresholdReference: l10n.t('Threshold: ≥ {pct}% of executions fan out (avg ≥ {fanout} partitions/query)', {
-            pct: thresholds.medPct,
-            fanout: thresholds.fanoutThreshold,
-        }),
-    };
-}
-
-export interface ShardKeyThresholds {
-    structuralPct: number;
-    minPartitions: number;
-    highPct: number;
-}
-
-/**
- * ShardKeyMisalignment (CODA DX-007): the structural escalation of DX-002. When the *dominant* share of query
- * executions fans out (≥ `structuralPct`) across ≥ `minPartitions` physical partitions, the problem is the
- * partition key itself (the data model), not individual queries — the fix is to re-key the container
- * (migration-class). Reuses DX-002's fan-out signal. When this fires it **supersedes** DX-002 on the same
- * container (the caller drops the DX-002 advisory). Container-scoped. Pure.
- */
-export function evaluateShardKeyMisalignment(
-    input: CrossPartitionInput,
-    thresholds: ShardKeyThresholds,
-    crossPartitionFanoutThreshold: number,
-): DerivedAdvisory | undefined {
-    const { containerPartitions, crossPartitionPct } = crossPartitionStats(input.shapes, crossPartitionFanoutThreshold);
-    if (crossPartitionPct < thresholds.structuralPct || containerPartitions < thresholds.minPartitions) {
-        return undefined;
-    }
-    const severity: DerivedAdvisorySeverity = crossPartitionPct >= thresholds.highPct ? 'High' : 'Medium';
-    const pct = Math.round(crossPartitionPct);
-    return {
-        id: `ShardKeyMisalignment/${containerKey(input.databaseId, input.containerId)}`,
-        rule: 'ShardKeyMisalignment',
-        severity,
-        scope: containerKey(input.databaseId, input.containerId),
-        title: l10n.t('Partition key misaligned in {container}', { container: input.containerId }),
-        rationale: clampRationale(
-            l10n.t(
-                '{pct}% of this container’s query executions fan out across ~{partitions} physical partitions — the partition key is structurally misaligned with the workload (the data model, not a single bad query).',
-                { pct, partitions: containerPartitions },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Re-key the container so the dominant query filter becomes the partition key. This is migration-class: create a new container with the corrected key, backfill the data, and cut over. Per-query fixes cannot resolve a fundamentally mismatched key, and the cost grows as the container adds partitions.',
-        ),
-        thresholdReference: l10n.t('Threshold: ≥ {pct}% of executions fan out across ≥ {partitions} partitions', {
-            pct: thresholds.structuralPct,
-            partitions: thresholds.minPartitions,
-        }),
-    };
-}
-
-/** Per-container write-dominance + throttling input (CODA DX-010). RU/request totals over the window. */
-export interface UncontrolledIngestionInput {
-    databaseId: string;
-    containerId: string;
-    /** RU consumed by write operations (Create/Upsert/Replace/Delete/Patch/Batch/Execute). */
-    writeRu: number;
-    /** Total RU consumed across all operations. */
-    totalRu: number;
-    /** Total data-plane requests in the window. */
-    totalRequests: number;
-    /** Requests that returned 429. */
-    throttledRequests: number;
-    /** Peak ÷ average per-minute write RU — supporting burstiness evidence (surfaced, not a gate). */
-    burstFactor: number;
-    /** Dominant write-path client `UserAgent`, surfaced as context when present. */
-    dominantUserAgent?: string;
-}
-
-export interface UncontrolledIngestionThresholds {
-    writeRuPctFloor: number;
-    throttleRatePct: number;
-    highPct: number;
-    minRequests: number;
-}
-
-/**
- * UncontrolledIngestion (CODA DX-010): a write-dominant workload that is throttling — the signature of batch /
- * bulk / Spark loads outrunning the container's throughput. Fires when there is enough traffic (`minRequests`),
- * the write-RU share is ≥ `writeRuPctFloor`, AND the 429 rate is ≥ `throttleRatePct` (High at/above `highPct`).
- * The fix is application-side rate control, complementary to DX-005's "add RU/s". Container-scoped. Pure.
- */
-export function evaluateUncontrolledIngestion(
-    input: UncontrolledIngestionInput,
-    thresholds: UncontrolledIngestionThresholds,
-): DerivedAdvisory | undefined {
-    if (input.totalRequests < thresholds.minRequests) {
-        return undefined;
-    }
-    const writeRuPct = input.totalRu > 0 ? (100 * input.writeRu) / input.totalRu : 0;
-    const throttleRate = input.totalRequests > 0 ? (100 * input.throttledRequests) / input.totalRequests : 0;
-    if (writeRuPct < thresholds.writeRuPctFloor || throttleRate < thresholds.throttleRatePct) {
-        return undefined;
-    }
-    const severity: DerivedAdvisorySeverity = throttleRate >= thresholds.highPct ? 'High' : 'Medium';
-    const writePct = Math.round(writeRuPct);
-    const throttlePct = Math.round(throttleRate);
-    const client = input.dominantUserAgent
-        ? l10n.t(' Dominant client: {client}.', { client: input.dominantUserAgent })
-        : '';
-    return {
-        id: `UncontrolledIngestion/${containerKey(input.databaseId, input.containerId)}`,
-        rule: 'UncontrolledIngestion',
-        severity,
-        scope: containerKey(input.databaseId, input.containerId),
-        title: l10n.t('Uncontrolled ingestion in {container}', { container: input.containerId }),
-        rationale: clampRationale(
-            l10n.t(
-                'This container is write-dominant ({writePct}% of RU on writes) and throttling at {throttlePct}% (peak-to-average write burst ~{burst}×), the signature of uncontrolled ingestion.',
-                { writePct, throttlePct, burst: input.burstFactor.toFixed(1) },
-            ) + client,
-        ),
-        suggestedAction: l10n.t(
-            'Add application-side rate control rather than only provisioning for the peak: bound ingestion concurrency and honor retry-after on 429; use the SDK throughput-control feature (or the Spark connector) or throughput buckets to cap the batch path; and/or run ingestion under low priority so it yields to app traffic. Raising RU/s is the complementary lever.',
-        ),
-        thresholdReference: l10n.t('Threshold: write-RU ≥ {writePct}% and 429 rate ≥ {throttlePct}%', {
-            writePct: thresholds.writeRuPctFloor,
-            throttlePct: thresholds.throttleRatePct,
-        }),
-    };
-}
-
-/** One collection's data-plane traffic within a shared-throughput database, over the window (CODA DX-003). */
-export interface CollectionTrafficInput {
-    containerId: string;
-    requests: number;
-    throttledRequests: number;
-    ruConsumed: number;
-}
-
-/** Per-shared-database input (CODA DX-003). */
-export interface SharedThroughputInput {
-    databaseId: string;
-    /** Provisioned RU/s shared across the database's collections, surfaced as context. */
-    sharedRu: number;
-    collections: readonly CollectionTrafficInput[];
-}
-
-export interface SharedThroughputThresholds {
-    minRequests: number;
-    poolThrottlePct: number;
-    dominancePct: number;
-    victimThrottlePct: number;
-    victimSharePct: number;
-}
-
-/**
- * SharedThroughputStarvation (CODA DX-003): in a shared-throughput database one collection monopolizes the
- * database-level RU pool, starving its siblings — they throttle despite consuming little while the pool is at
- * its cap. Fires when ≥ 2 collections are active, the pool 429 rate is ≥ `poolThrottlePct`, one collection
- * consumes ≥ `dominancePct` of the pool RU, and a different collection throttles ≥ `victimThrottlePct` while
- * consuming ≤ `victimSharePct`. The internal per-collection RU split is not exposed, so this infers starvation
- * from the 429 + consumption disparity. Database-scoped, High severity. Pure.
- */
-export function evaluateSharedThroughputStarvation(
-    input: SharedThroughputInput,
-    thresholds: SharedThroughputThresholds,
-): DerivedAdvisory | undefined {
-    const active = input.collections.filter((c) => c.requests >= thresholds.minRequests);
-    if (active.length < 2) {
-        return undefined;
-    }
-    const ruTotal = active.reduce((sum, c) => sum + c.ruConsumed, 0);
-    const reqTotal = active.reduce((sum, c) => sum + c.requests, 0);
-    const thrTotal = active.reduce((sum, c) => sum + c.throttledRequests, 0);
-    const poolRate = reqTotal > 0 ? (100 * thrTotal) / reqTotal : 0;
-    if (poolRate < thresholds.poolThrottlePct || ruTotal <= 0) {
-        return undefined;
-    }
-    const share = (c: CollectionTrafficInput): number => (100 * c.ruConsumed) / ruTotal;
-    const throttleRate = (c: CollectionTrafficInput): number =>
-        c.requests > 0 ? (100 * c.throttledRequests) / c.requests : 0;
-    const monopolizer = active.reduce((top, c) => (share(c) > share(top) ? c : top), active[0]);
-    const monoShare = share(monopolizer);
-    const victims = active.filter(
-        (c) =>
-            c.containerId !== monopolizer.containerId &&
-            throttleRate(c) >= thresholds.victimThrottlePct &&
-            share(c) <= thresholds.victimSharePct,
-    );
-    if (monoShare < thresholds.dominancePct || victims.length === 0) {
-        return undefined;
-    }
-    const victim = victims.reduce((worst, c) => (throttleRate(c) > throttleRate(worst) ? c : worst), victims[0]);
-    return {
-        id: `SharedThroughputStarvation/${input.databaseId}`,
-        rule: 'SharedThroughputStarvation',
-        severity: 'High',
-        scope: input.databaseId,
-        title: l10n.t('Shared-throughput starvation in {database}', { database: input.databaseId }),
-        rationale: clampRationale(
-            l10n.t(
-                'In shared-throughput database {database} ({ru} RU/s across {count} collections), {mono} consumes {monoShare}% of the pool while {victim} throttles at {victimRate}% despite using only {victimShare}% — it is being starved.',
-                {
-                    database: input.databaseId,
-                    ru: Math.round(input.sharedRu),
-                    count: active.length,
-                    mono: monopolizer.containerId,
-                    monoShare: Math.round(monoShare),
-                    victim: victim.containerId,
-                    victimRate: Math.round(throttleRate(victim)),
-                    victimShare: Math.round(share(victim)),
-                },
-            ),
-        ),
-        suggestedAction: l10n.t(
-            'Move the monopolizing collection (or the starved one) to dedicated throughput, or migrate collections off the shared pool. The exact internal RU split per collection is not exposed; this is inferred from the 429 + consumption disparity.',
-        ),
-        thresholdReference: l10n.t(
-            'Threshold: pool 429 ≥ {pool}%, one collection ≥ {dominance}% of RU, a sibling throttling ≥ {victim}% at ≤ {share}% share',
-            {
-                pool: thresholds.poolThrottlePct,
-                dominance: thresholds.dominancePct,
-                victim: thresholds.victimThrottlePct,
-                share: thresholds.victimSharePct,
-            },
-        ),
-    };
-}
-
-/** Already-fetched telemetry the derived-advisory engine consumes. */
-export interface DerivedAdvisoryInputs {
-    /** Account-wide 7-day RU% samples for the over-provisioning + autoscale rules. */
-    weeklyRuPercents: readonly number[];
-    /** Max of `weeklyRuPercents`, or undefined when no 7-day data is available. */
-    weeklyPeakPercent?: number;
-    /** True when at least one container/database uses manual (non-autoscale, non-serverless) throughput. */
-    hasManualThroughput: boolean;
-    /** Total provisioned RU/s across manual-throughput containers, for over-provisioning materiality. */
-    manualProvisionedRuTotal?: number;
-    /** Per-container physical-partition p99 saturation + 429 rate for the hot-partition and under-provisioning rules. */
-    partitions: readonly PartitionSaturationInput[];
-    /** Per-container physical-partition storage series (7d) for the storage-growth and storage-skew rules. */
-    storage: readonly ContainerStorageInput[];
-    /** Per-container index/data storage for the indexing-cost rule. */
-    indexing: readonly IndexingUsageInput[];
-    /** Per-container idle signal (peak `TotalRequestUnits` per bucket over 30d) for the idle rule (DX-004). */
-    idleContainers?: readonly IdleContainerInput[];
-    /** Per-container physical-partition count + storage for the partition-merge rule (DX-009). */
-    partitionMerges?: readonly PartitionMergeInput[];
-    /** Per-autoscale-container `AutoscaledRU`-%-of-max duty cycle for the autoscale rules (DX-011 / DX-013). */
-    autoscaleUtilizations?: readonly AutoscaleUtilizationInput[];
-    /** Account-total consumed-RU shape (30d) for the serverless-candidate rule (DX-014); absent when unavailable. */
-    serverless?: ServerlessCandidateInput;
-    /** Total provisioned RU/s across the account, the materiality denominator for the idle/autoscale-max rules. */
-    scopeProvisionedRuTotal?: number;
-    /** Account-level ARM configuration for the config-only advisories (DX-016 / DX-008); absent when unavailable. */
-    accountConfig?: AccountConfigInput;
-    /** Per-container query fan-out shapes (Log Analytics) for the cross-partition + shard-key rules (DX-002 / DX-007). */
-    crossPartition?: readonly CrossPartitionInput[];
-    /** Per-container write-dominance + throttling (Log Analytics) for the uncontrolled-ingestion rule (DX-010). */
-    ingestion?: readonly UncontrolledIngestionInput[];
-    /** Per-shared-database collection traffic (Log Analytics) for the shared-throughput-starvation rule (DX-003). */
-    sharedThroughput?: readonly SharedThroughputInput[];
-}
-
-/**
- * Runs every derived-advisory rule over already-fetched telemetry and returns the advisories
- * that fired, sorted by severity. Pure so the whole engine is unit-testable.
- */
-export function computeDerivedAdvisories(
-    inputs: DerivedAdvisoryInputs,
-    thresholds: DerivedAdvisoryThresholds,
-): DerivedAdvisory[] {
-    const advisories: DerivedAdvisory[] = [];
-
-    for (const partition of inputs.partitions) {
-        const hot = evaluateHotPartitionRisk(
-            partition,
-            thresholds.partitionSaturationPercent,
-            thresholds.partitionHeadroomPercent,
-            thresholds.throttleRatePercent,
-        );
-        if (hot) {
-            advisories.push(hot);
-        }
-        // DX-005 only fires on uniform saturation; the skew case is already covered by HotPartitionRisk above,
-        // so the two never double-report the same throttling.
-        const under = evaluateUnderProvisioning(
-            partition,
-            thresholds.partitionSaturationPercent,
-            thresholds.partitionHeadroomPercent,
-            thresholds.throttleRatePercent,
-        );
-        if (under) {
-            advisories.push(under);
-        }
-    }
-
-    const overProvisioning = evaluateOverProvisioning(
-        percentile(inputs.weeklyRuPercents, 99),
-        inputs.weeklyPeakPercent ??
-            (inputs.weeklyRuPercents.length > 0 ? Math.max(...inputs.weeklyRuPercents) : undefined),
-        thresholds.overProvisioningBandPercent,
-        inputs.hasManualThroughput,
-        inputs.manualProvisionedRuTotal,
-    );
-    if (overProvisioning) {
-        advisories.push(overProvisioning);
-    }
-
-    const autoscale = evaluateAutoscaleCandidate(
-        inputs.weeklyPeakPercent ??
-            (inputs.weeklyRuPercents.length > 0 ? Math.max(...inputs.weeklyRuPercents) : undefined),
-        inputs.weeklyRuPercents.length > 0 ? mean(inputs.weeklyRuPercents) : undefined,
-        {
-            maxPercent: thresholds.autoscaleMaxPercent,
-            avgPercent: thresholds.autoscaleAvgPercent,
-            peakToAvgRatio: thresholds.autoscalePeakToAvgRatio,
-        },
-        inputs.hasManualThroughput,
-    );
-    if (autoscale) {
-        advisories.push(autoscale);
-    }
-
-    for (const storage of inputs.storage) {
-        const growth = evaluateStorageGrowthRisk(storage, thresholds.storageGrowthHorizonDays);
-        if (growth) {
-            advisories.push(growth);
-        }
-        const skew = evaluateStorageSkewRisk(storage, thresholds.storageSkewBalanceRatio);
-        if (skew) {
-            advisories.push(skew);
-        }
-    }
-
-    for (const indexing of inputs.indexing) {
-        const advisory = evaluateIndexingCostRisk(indexing, thresholds.indexingUsageRatio);
-        if (advisory) {
-            advisories.push(advisory);
-        }
-    }
-
-    for (const idle of inputs.idleContainers ?? []) {
-        const advisory = evaluateIdleContainer(idle, thresholds.idlePeakRuPerBucket, inputs.scopeProvisionedRuTotal);
-        if (advisory) {
-            advisories.push(advisory);
-        }
-    }
-
-    for (const merge of inputs.partitionMerges ?? []) {
-        const advisory = evaluatePartitionMergeCandidate(merge);
-        if (advisory) {
-            advisories.push(advisory);
-        }
-    }
-
-    for (const util of inputs.autoscaleUtilizations ?? []) {
-        const overMax = evaluateAutoscaleMaxOverProvisioned(
-            util,
-            thresholds.overProvisioningBandPercent,
-            inputs.scopeProvisionedRuTotal,
-        );
-        if (overMax) {
-            advisories.push(overMax);
-        }
-        const toManual = evaluateAutoscaleToManualCandidate(
-            util,
-            thresholds.autoscaleToManualAvgPercent,
-            thresholds.autoscaleToManualPeakToAvgRatio,
-        );
-        if (toManual) {
-            advisories.push(toManual);
-        }
-    }
-
-    if (inputs.serverless) {
-        const serverless = evaluateServerlessCandidate(inputs.serverless, {
-            sporadicRatio: thresholds.serverlessSporadicRatio,
-            peakFloorRuPerSec: thresholds.serverlessPeakFloorRuPerSec,
-            peakCeilingRuPerSec: thresholds.serverlessPeakCeilingRuPerSec,
-        });
-        if (serverless) {
-            advisories.push(serverless);
-        }
-    }
-
-    if (inputs.accountConfig) {
-        const consistency = evaluateExpensiveConsistency(inputs.accountConfig);
-        if (consistency) {
-            advisories.push(consistency);
-        }
-        const multiRegionWrites = evaluateMultiRegionWrites(inputs.accountConfig);
-        if (multiRegionWrites) {
-            advisories.push(multiRegionWrites);
-        }
-    }
-
-    // DX-002 / DX-007 (Log Analytics): DX-007 (structural re-key) supersedes DX-002 (per-query fix) on the same
-    // container — when the shard key is misaligned, per-query fixes cannot help — so emit only one per container.
-    for (const input of inputs.crossPartition ?? []) {
-        const shardKey = evaluateShardKeyMisalignment(
-            input,
-            {
-                structuralPct: thresholds.shardKeyStructuralPct,
-                minPartitions: 2,
-                highPct: thresholds.shardKeyHighPct,
-            },
-            thresholds.crossPartitionFanoutThreshold,
-        );
-        if (shardKey) {
-            advisories.push(shardKey);
-            continue;
-        }
-        const crossPartition = evaluateCrossPartitionQuery(input, {
-            minQueries: thresholds.crossPartitionMinQueries,
-            fanoutThreshold: thresholds.crossPartitionFanoutThreshold,
-            highPct: thresholds.crossPartitionHighPct,
-            medPct: thresholds.crossPartitionMedPct,
-        });
-        if (crossPartition) {
-            advisories.push(crossPartition);
-        }
-    }
-
-    // DX-010 (Log Analytics): write-dominant throttling per container.
-    for (const input of inputs.ingestion ?? []) {
-        const advisory = evaluateUncontrolledIngestion(input, {
-            writeRuPctFloor: thresholds.ingestionWriteRuPctFloor,
-            throttleRatePct: thresholds.ingestionThrottleRatePct,
-            highPct: thresholds.ingestionHighPct,
-            minRequests: thresholds.ingestionMinRequests,
-        });
-        if (advisory) {
-            advisories.push(advisory);
-        }
-    }
-
-    // DX-003 (Log Analytics): shared-throughput starvation per shared database.
-    for (const input of inputs.sharedThroughput ?? []) {
-        const advisory = evaluateSharedThroughputStarvation(input, {
-            minRequests: thresholds.sharedThroughputMinRequests,
-            poolThrottlePct: thresholds.sharedThroughputPoolThrottlePct,
-            dominancePct: thresholds.sharedThroughputDominancePct,
-            victimThrottlePct: thresholds.sharedThroughputVictimThrottlePct,
-            victimSharePct: thresholds.sharedThroughputVictimSharePct,
-        });
-        if (advisory) {
-            advisories.push(advisory);
-        }
-    }
-
-    advisories.sort(compareAdvisories);
-    return advisories;
-}
-
-// ─── Server-side orchestration ──────────────────────────────────────────────────
-
-/** Manual (non-autoscale, non-serverless) throughput modes that the over-provisioning/autoscale rules key on. */
-const MANUAL_THROUGHPUT_MODES: ReadonlySet<ThroughputMode> = new Set<ThroughputMode>(['dedicated', 'shared']);
-
-/** Upper bound on per-container partition-health scans per derived-advisories tick, to cap ARM reads. */
-const MAX_PARTITION_SCAN = 10;
-
-/** Everything {@link collectDerivedAdvisories} needs, gathered by the router as data preparation. */
-export interface CollectDerivedAdvisoriesParams {
-    monitorClient: MonitorClient;
-    cosmosClient: CosmosDBManagementClient;
-    accountId: string;
-    resourceGroup: string;
-    accountName: string;
-    isServerless: boolean;
-    healthThresholds: HealthThresholds;
-    partitionThresholds: PartitionThresholds;
-    advisoryThresholds: DerivedAdvisoryThresholds;
-    /** Account-level ARM configuration for the config-only advisories (DX-016 / DX-008). */
-    accountConfig?: AccountConfigInput;
-    /**
-     * Data-plane Log Analytics client for the Tier-2 log-based advisories (DX-002/003/007/010). Absent when the
-     * client could not be acquired — the collector then reports Tier-2 as unavailable (`noData`) while still
-     * returning the Tier-1 advisories.
-     */
-    logsClient?: LogsQueryExecutor;
-}
-
-/** What {@link collectDerivedAdvisories} returns: the fired advisories plus the Tier-2 (Log Analytics) coverage status. */
-export interface CollectedDerivedAdvisories {
-    advisories: DerivedAdvisory[];
-    logSource: LogsSourceStatus;
-}
-
-/**
- * Gathers the telemetry the derived-advisory rules need — static inventory, 24h
- * and 7d RU trends, 7d inventory metrics, and per-container partition health for
- * the busiest containers — then runs {@link computeDerivedAdvisories}. Any Azure
- * failure propagates so the router can classify it into an empty-state reason.
- */
-export async function collectDerivedAdvisories(
-    params: CollectDerivedAdvisoriesParams,
-): Promise<CollectedDerivedAdvisories> {
-    const {
-        monitorClient,
-        cosmosClient,
-        accountId,
-        resourceGroup,
-        accountName,
-        isServerless,
-        healthThresholds,
-        partitionThresholds,
-        advisoryThresholds,
-        accountConfig,
-        logsClient,
-    } = params;
-
-    const rows = await getSqlInventory(cosmosClient, resourceGroup, accountName, isServerless);
-
-    const [weekTrends, inventoryMetrics, idlePeaks, partitionCounts, serverlessShape] = await Promise.all([
-        getRuTrends(monitorClient, accountId, '7D', undefined, undefined),
-        getInventoryMetrics(monitorClient, accountId, '7D', undefined, healthThresholds),
-        // DX-004: peak consumed RU per bucket over 30 days, per container (empty ⇒ idle, not an error).
-        getContainerIdlePeaks(monitorClient, accountId).catch(() => new Map<string, number>()),
-        // DX-009: actual physical-partition count per container.
-        getContainerPartitionCounts(monitorClient, accountId).catch(() => new Map<string, number>()),
-        // DX-014: account-total consumed-RU shape over 30 days (skipped for accounts already on serverless).
-        isServerless
-            ? Promise.resolve(undefined)
-            : getAccountConsumedRuShape(monitorClient, accountId).catch(() => undefined),
-    ]);
-
-    // Scan partition health only for the busiest containers (highest 7-day peak RU) to bound ARM reads.
-    const scanTargets = [...rows]
-        .sort(
-            (a, b) =>
-                (inventoryMetrics.metrics[containerKey(b.databaseId, b.containerId)]?.peakRuPercent ?? -1) -
-                (inventoryMetrics.metrics[containerKey(a.databaseId, a.containerId)]?.peakRuPercent ?? -1),
-        )
-        .slice(0, MAX_PARTITION_SCAN);
-
-    // DX-006 / DX-005 are present-tense over a 7-day window: per-physical-partition p99 saturation plus the
-    // container's aggregate 429 rate, fetched together for each scanned container.
-    const partitionResults = await Promise.all(
-        scanTargets.map((row) =>
-            getPartitionHealth(
-                monitorClient,
-                accountId,
-                'ru',
-                '7D',
-                row.databaseId,
-                row.containerId,
-                partitionThresholds,
-            ).catch(() => undefined),
-        ),
-    );
-
-    const throttleResults = await Promise.all(
-        scanTargets.map((row) =>
-            getThrottleRate(monitorClient, accountId, '7D', row.databaseId, row.containerId).catch(() => undefined),
-        ),
-    );
-
-    const storageResults = await Promise.all(
-        scanTargets.map((row) =>
-            getPartitionStorageSeries(monitorClient, accountId, '7D', row.databaseId, row.containerId).catch(
-                () => undefined,
-            ),
-        ),
-    );
-
-    // DX-011 / DX-013: the AutoscaledRU-%-of-max duty cycle, fetched only for the scanned autoscale containers
-    // (each is a per-container query, so bound it to the same busiest-container set as the partition scans).
-    const autoscaleTargets = scanTargets.filter(
-        (row) => row.throughputMode === 'autoscale' && row.throughputRU !== undefined && row.throughputRU > 0,
-    );
-    const autoscaleResults = await Promise.all(
-        autoscaleTargets.map((row) =>
-            getAutoscaleUtilization(monitorClient, accountId, row.databaseId, row.containerId, row.throughputRU!).catch(
-                () => undefined,
-            ),
-        ),
-    );
-
-    const rowByKey = new Map(rows.map((row) => [containerKey(row.databaseId, row.containerId), row]));
-    const partitions: PartitionSaturationInput[] = [];
-    partitionResults.forEach((result, index) => {
-        if (!result || !result.available || result.mode !== 'ru') {
-            return;
-        }
-        const row = rowByKey.get(containerKey(result.databaseId, result.containerId));
-        const throttle = throttleResults[index];
-        partitions.push({
-            databaseId: result.databaseId,
-            containerId: result.containerId,
-            maxP99: result.maxSaturationPercent ?? result.topPartitionShare,
-            minP99: result.minSaturationPercent ?? 0,
-            partitionCount: result.partitionCount,
-            throttleRatePercent: throttle?.ratePercent ?? 0,
-            totalRequests: throttle?.totalRequests ?? 0,
-            throughputMode: row?.throughputMode ?? 'unknown',
-            provisionedRu: row?.throughputRU,
-        });
-    });
-
-    const storage = storageResults
-        .filter((result): result is PartitionStorageResult => !!result && result.available)
-        .map((result) => ({
-            databaseId: result.databaseId,
-            containerId: result.containerId,
-            partitions: result.partitions,
-        }));
-
-    const indexing: IndexingUsageInput[] = rows.map((row) => {
-        const metrics = inventoryMetrics.metrics[containerKey(row.databaseId, row.containerId)];
-        return {
-            databaseId: row.databaseId,
-            containerId: row.containerId,
-            indexUsageBytes: metrics?.indexUsageBytes,
-            dataUsageBytes: metrics?.dataUsageBytes,
-            excludedPathCount: row.excludedPathCount,
-        };
-    });
-
-    const weeklyRuPercents = weekTrends.points
-        .map((point) => point.ruPercent)
-        .filter((value): value is number => value !== undefined);
-
-    const manualProvisionedRuTotal = rows
-        .filter((row) => MANUAL_THROUGHPUT_MODES.has(row.throughputMode) && row.throughputRU !== undefined)
-        .reduce<number | undefined>((sum, row) => (sum ?? 0) + (row.throughputRU ?? 0), undefined);
-
-    // DX-004: build an idle input for every container with a recoverable offer (manual or autoscale). A container
-    // absent from the peaks map never emitted RU over the window, which is the idle signal (peak defaults to 0).
-    const idleContainers: IdleContainerInput[] = rows
-        .filter((row) => row.throughputRU !== undefined && row.throughputMode !== 'serverless')
-        .map((row) => ({
-            databaseId: row.databaseId,
-            containerId: row.containerId,
-            peakRuPerBucket: idlePeaks.get(containerKey(row.databaseId, row.containerId)) ?? 0,
-            throughputMode: row.throughputMode,
-            provisionedRu: row.throughputRU,
-        }));
-
-    // DX-009: pair each container's actual physical-partition count with its provisioned RU/s and current storage.
-    const partitionMerges: PartitionMergeInput[] = rows.map((row) => {
-        const metrics = inventoryMetrics.metrics[containerKey(row.databaseId, row.containerId)];
-        return {
-            databaseId: row.databaseId,
-            containerId: row.containerId,
-            actualPartitions: partitionCounts.get(containerKey(row.databaseId, row.containerId)) ?? 0,
-            provisionedRu: row.throughputRU,
-            dataUsageBytes: metrics?.dataUsageBytes,
-        };
-    });
-
-    // DX-011 / DX-013: the fetched autoscale duty cycles, keyed back to their containers' configured max.
-    const autoscaleUtilizations: AutoscaleUtilizationInput[] = [];
-    autoscaleResults.forEach((result, index) => {
-        if (!result) {
-            return;
-        }
-        const row = autoscaleTargets[index];
-        autoscaleUtilizations.push({
-            databaseId: row.databaseId,
-            containerId: row.containerId,
-            peakPercent: result.peakPercent,
-            avgPercent: result.avgPercent,
-            sampleCount: result.sampleCount,
-            configuredMaxRu: row.throughputRU,
-        });
-    });
-
-    // DX-014: account-total consumed-RU shape (absent for serverless accounts or when the fetch failed).
-    const serverless: ServerlessCandidateInput | undefined = serverlessShape
-        ? { ...serverlessShape, isServerless }
-        : undefined;
-
-    // Materiality denominator (DX-004 / DX-011): total provisioned RU/s across the whole account.
-    const scopeProvisionedRuTotal = rows.reduce<number | undefined>(
-        (sum, row) => (row.throughputRU === undefined ? sum : (sum ?? 0) + row.throughputRU),
-        undefined,
-    );
-
-    // Tier-2 (Log Analytics): DX-002/007 (query fan-out) and DX-010 (write-dominant throttling) for the busiest
-    // scanned containers, and DX-003 (shared-throughput starvation) per shared-throughput database. Bounded to the
-    // same busiest-container set as the ARM scans. Probed once so a single reason-specific partial-coverage notice
-    // is surfaced when logs are unavailable, and Tier-1 still ships.
-    const logs = await collectTier2Logs(logsClient, accountId, scanTargets, rows);
-
-    const advisories = computeDerivedAdvisories(
-        {
-            weeklyRuPercents,
-            weeklyPeakPercent: weekTrends.peakPercent,
-            hasManualThroughput: rows.some((row) => MANUAL_THROUGHPUT_MODES.has(row.throughputMode)),
-            manualProvisionedRuTotal,
-            partitions,
-            storage,
-            indexing,
-            idleContainers,
-            partitionMerges,
-            autoscaleUtilizations,
-            serverless,
-            scopeProvisionedRuTotal,
-            accountConfig,
-            crossPartition: logs.crossPartition,
-            ingestion: logs.ingestion,
-            sharedThroughput: logs.sharedThroughput,
-        },
-        advisoryThresholds,
-    );
-
-    return { advisories, logSource: logs.logSource };
-}
-
-/** The Tier-2 inputs gathered from Log Analytics plus the coverage status the card surfaces. */
-interface Tier2Logs {
-    crossPartition: CrossPartitionInput[];
-    ingestion: UncontrolledIngestionInput[];
-    sharedThroughput: SharedThroughputInput[];
-    logSource: LogsSourceStatus;
-}
-
-/**
- * Fetches the Tier-2 (Log Analytics) inputs for the busiest scanned containers and shared-throughput databases,
- * degrading gracefully: with no client Tier-2 is reported unavailable (`noData`); if the `CDB*` tables are not
- * queryable the probe classifies why (`logAnalyticsDisabled` / `rbac` / `noData`) and no per-detector queries run.
- * Individual query failures degrade that detector to empty rather than failing the whole section.
- */
-async function collectTier2Logs(
-    logsClient: LogsQueryExecutor | undefined,
-    accountId: string,
-    scanTargets: readonly { databaseId: string; containerId: string }[],
-    rows: readonly { databaseId: string; containerId: string; throughputMode: ThroughputMode; throughputRU?: number }[],
-): Promise<Tier2Logs> {
-    const empty = { crossPartition: [], ingestion: [], sharedThroughput: [] };
-    if (!logsClient) {
-        return { ...empty, logSource: { available: false, reason: 'noData' } };
-    }
-
-    const timespan = logWindow();
-    const probe = await probeCdbLogs(logsClient, accountId, timespan);
-    if (!probe.available) {
-        return { ...empty, logSource: { available: false, reason: probe.reason } };
-    }
-
-    // DX-003: one query per shared-throughput database (deduped), using the shared pool's provisioned RU/s.
-    const sharedDatabases = new Map<string, number>();
-    for (const row of rows) {
-        if (row.throughputMode === 'shared' && !sharedDatabases.has(row.databaseId)) {
-            sharedDatabases.set(row.databaseId, row.throughputRU ?? 0);
-        }
-    }
-
-    const [crossPartitionResults, ingestionResults, sharedResults] = await Promise.all([
-        Promise.all(
-            scanTargets.map((row) =>
-                fetchCrossPartitionShapes(logsClient, accountId, row.databaseId, row.containerId, timespan).catch(
-                    () => undefined,
-                ),
-            ),
-        ),
-        Promise.all(
-            scanTargets.map((row) =>
-                fetchUncontrolledIngestion(logsClient, accountId, row.databaseId, row.containerId, timespan).catch(
-                    () => undefined,
-                ),
-            ),
-        ),
-        Promise.all(
-            [...sharedDatabases].map(([databaseId, sharedRu]) =>
-                fetchSharedThroughputTraffic(logsClient, accountId, databaseId, sharedRu, timespan).catch(
-                    () => undefined,
-                ),
-            ),
-        ),
-    ]);
-
-    return {
-        crossPartition: crossPartitionResults.filter((r): r is CrossPartitionInput => r !== undefined),
-        ingestion: ingestionResults.filter((r): r is UncontrolledIngestionInput => r !== undefined),
-        sharedThroughput: sharedResults.filter((r): r is SharedThroughputInput => r !== undefined),
-        logSource: { available: true },
-    };
-}
